Remove unused x-SDCH header from get requests
This was originally in place to allow the server folks to filter out (disable)
a specific client.  Given that it is no big deal if chrome support for SDCH 
is disabled in GWS, this is really not needed (and it accounts for a large
number of bytes in ALL get requests.

If we have a problem with SDCH support in chrome, we'd roll a new version

IMO, if server needs to disable this feature for a specific version, they
should filter on the actual user agent string, and not require a redundant 
copy of it here.

r=openvcdiff,huanr,kmixter
Review URL: http://codereview.chromium.org/13071

git-svn-id: http://src.chromium.org/svn/trunk/src/net@6259 4ff67af0-8c30-449e-8e8b-ad334ec8d88c