Third time's a charm? Fix part of http://code.google.com/p/chromium/issues/detail?id=4510 and improve https support in test shell on linux.
This is the same as the earlier ssl cert cl
(see http://codereview.chromium.org/11249),
but with the certs moved so net can use them without
reaching over into chrome's pants and causing
test failure on the 'modules' Windows build server,
which is set up to test net and base but not chrome.

For this to pass, we will need to install 
the certs on the windows module and try servers.
(And make sure tlslite is present.)

(A later CL will finish implementing SSLInfo for Linux,
and probably reference net/base/ssl_test_util.cc 
from all three vcproj files that need it, 
even though that's ugly, because that's less ugly
that referencing it from net.lib's vcproj.)


Review URL: http://codereview.chromium.org/12930

git-svn-id: http://src.chromium.org/svn/trunk/src/net@6495 4ff67af0-8c30-449e-8e8b-ad334ec8d88c