Make sure filter buffer is flushed after it fills its output quota
In some cases, when the filter returned but the output buffer was full,
the ReadFilteredData code was not "bothering" to call the filter again
(because there was no more input data to provide).  We now maintain
state indicating that the output buffer was filled (last time) and use
that to ensure we call the filter(s) again when Read is next called.

bug=7985

r=huanr
Review URL: http://codereview.chromium.org/27073

git-svn-id: http://src.chromium.org/svn/trunk/src/net@10244 4ff67af0-8c30-449e-8e8b-ad334ec8d88c