Make sure filters can handle an empty input buffer
The chaining of filters helped to guarantee that buffers were properly
flushed when large expansions took place.  That change assumed that a filter
could be called with no input (as some filters need to be called
repeatedly this way to purge massive internal state).  The gzip and
bzip filters were returning an error in that scenario.

bug=9316

r=huanr
Review URL: http://codereview.chromium.org/53074

git-svn-id: http://src.chromium.org/svn/trunk/src/net@12524 4ff67af0-8c30-449e-8e8b-ad334ec8d88c