Local text file with spaces in filename is urlencoded in tab title
When viewing a local text file with spaces in filename, it is still urlencoded. Filename should be displayed with spaces, not with urlencoding. It would be more user-friendly.

Since net::FormatURL is already implemented, using it would be great. But it doesn't escape SPACES, just NORMAL, it doesn't even escape unicode. I plumbed out a unescapeurl that could be used whether we allow conversion of spaces or not.

BUG=8775 (http://crbug.com/8775)
TEST=Tested whether the input is escaped in the navigational context and ran the net tests

New Review: http://codereview.chromium.org/118059
Review URL: http://codereview.chromium.org/56053

git-svn-id: http://src.chromium.org/svn/trunk/src/net@17462 4ff67af0-8c30-449e-8e8b-ad334ec8d88c