Fix a browser crash during cache validation.
It is possible for a crazy server to return 206 even
though we are not making a range request. In that
case, we should not allow the stored response to go
through the logic for range support (which right now
is disabled). We've always stored those 206s though.

BUG=15617
TEST=unittest

Review URL: http://codereview.chromium.org/150090

git-svn-id: http://src.chromium.org/svn/trunk/src/net@19621 4ff67af0-8c30-449e-8e8b-ad334ec8d88c