Disk cache: Keep a map of all open entries.
We still have a few crashes when for some reason we believe
an entry is not dirty and we follow the pointer stored by
its rankings node only to crash while accessing the memory.
I have no explanation to why the dirty id matches the current
one (a page boundary issue maybe?), but having a map with all
open entries solves the issue of having to follow pointers
from disk.

BUG=15596, b/1120346
TEST=unittests

Review URL: http://codereview.chromium.org/149218

git-svn-id: http://src.chromium.org/svn/trunk/src/net@20067 4ff67af0-8c30-449e-8e8b-ad334ec8d88c