Disk cache: Add explicit support for eviction / deletion of sparse entries.
I started to add code to modify the children_map of the
parent entry when a child is evicted but that ended up
being too much trouble for too little gain. We have to be
prepared to handle the case of not finding a child entry
because there is no way to make sure that the process
doesn't go away at any time, so adding a lot of complexity
just to avoid an extra entry lookup is just not worth it.

On the other hand, potentially freeing up a lot of space when
a sparse entry is deleted (insetad of just waiting for the
eviction code to do the cleanup) seems like a good thing.

BUG=12258
TEST=unittest

Review URL: http://codereview.chromium.org/149306

git-svn-id: http://src.chromium.org/svn/trunk/src/net@20325 4ff67af0-8c30-449e-8e8b-ad334ec8d88c