Disk Cache: Don't depend on the backend being enabled to be able to return the key of an open entry.
Whenever a critical corruption is detected by the disk cache,
the backend disables itself and starts failing all requests
until it's able to re-create the backing store.

Key's longer than 928 bytes are not stored inside the entry
itself, so a file object is required to access them. The backend
will reject any request for a file object after it is disabled,
so a user's request for the key of an open entry will also
fail.

Now we keep a pointer to the related file object (if needed) so
that we don't have to ask the backend for it when the user
requests the current key.

BUG=9952
TEST=unittest


Review URL: http://codereview.chromium.org/165030

git-svn-id: http://src.chromium.org/svn/trunk/src/net@22637 4ff67af0-8c30-449e-8e8b-ad334ec8d88c