ForceTLS: persist to disk
With this patch, we'll persist ForceTLS state to disk. It's saved as a
JSON file (ForceTLSState) in the user data directory for the moment.
You still need the --force-https flag in order to trigger any ForceTLS
behaviour, however.

For the moment, this state isn't cleared when the rest of the browser
state it. That's ok because it's still behind a flag.

http://codereview.chromium.org/186014


git-svn-id: http://src.chromium.org/svn/trunk/src/net@25382 4ff67af0-8c30-449e-8e8b-ad334ec8d88c