ForceTLS: generate internal redirect when mistakenly using http.
Previously, we just generated a cryptic error when the user attempted
to navigate to a ForceTLS site with HTTP. With this patch, we generate
an internal redirect to the HTTPS version of the site.

(This collided with Dave's r25496, which also added
url_request_redirect_job.cc. Thankfully our code was almost identical.)

http://codereview.chromium.org/193014


git-svn-id: http://src.chromium.org/svn/trunk/src/net@25519 4ff67af0-8c30-449e-8e8b-ad334ec8d88c