Re-enabled URLRequestTestHTTP.BasicAuth.
The problem was the test was expecting |response_time| of cached responses to be the timestamp of the original request, rather than the timestamp of the last cache validation.

Also updated the header to clarify what |request_time| and |response_time| correspond to in the cached cases.

BUG=http://crbug.com/20594

Review URL: http://codereview.chromium.org/197030

git-svn-id: http://src.chromium.org/svn/trunk/src/net@25540 4ff67af0-8c30-449e-8e8b-ad334ec8d88c