Strict transport security: add checkbox to clear state.
This patches add a checkbox to the "Clear Browsing Data" dialog which
clears the STS state when checked. Since we don't timestamp our
entries (for now at least, should we?), the duration selected has no
effect: we always clear everything.

Mac doesn't appear to have a dialog for this yet, so no Mac specific
changes.

http://codereview.chromium.org/196070


git-svn-id: http://src.chromium.org/svn/trunk/src/net@25955 4ff67af0-8c30-449e-8e8b-ad334ec8d88c