Fix two issues with a downloaded file's extension:
1. On Mac, we were prepending an extra '.' to the extension causing
it to fail a comparison (resulting in second extension being added).
2. For downloads with the extension '.tgz' and mime type
'application/x-gzip', we were appending an extra '.gz' extension.

The extension compare code is starting to get a little long, but I
don't think it's too bad if there are only a few of these cases.

BUG=21591,21595
TEST=1. Download a file on Mac with a well known extension type (mp3, gz)
and notice that there are no double extensions. 2. Download a tgz file and
notice that the file's extension is '.tgz', not '.tgz.gz'.

Review URL: http://codereview.chromium.org/195064

git-svn-id: http://src.chromium.org/svn/trunk/src/net@26028 4ff67af0-8c30-449e-8e8b-ad334ec8d88c