Implement ScopedRunnableMethodFactory using WeakPtr.
This required some changes to WeakPtr to support the addition
of WeakPtrFactory::HasWeakPtrs(), which is used to implement
ScopedRunnableMethodFactory::empty().

Now, the WeakReferenceOwner just holds a pointer to the Flag
class, and the Flag holds a back-pointer that it can use to
clear the WeakReferenceOwner's pointer when the Flag is
destroyed.  I use the null'ness of this back-pointer in place
of the bool member that was previously used to indicate if the
WeakReference is valid.

It was also necessary to expose a HasOneRef method on
RefCounted.  I included one on RefCountedThreadSafe for
completeness.

Finally, I switched HttpCache over to using WeakPtr instead
of RevocableStore so that I could delete RevocableStore.

(I'm making this change to consolidate similar functionality.)

R=abarth
BUG=none
TEST=none

Review URL: http://codereview.chromium.org/235027

git-svn-id: http://src.chromium.org/svn/trunk/src/net@27287 4ff67af0-8c30-449e-8e8b-ad334ec8d88c