Fix two flip bugs.
First, we kept points to FlipDelegates rather than references.
Now that the classes are separated, this becomes problematic
because the FlipNetworkTransaction can go away independently 
of the FlipSession.

Second, there was a bug with posted data - we were compressing
the frames, and we don't want to.

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/402073

git-svn-id: http://src.chromium.org/svn/trunk/src/net@32458 4ff67af0-8c30-449e-8e8b-ad334ec8d88c