Revert 38001 and 38002.
Modify CookiePolicy to work asynchronously

This change will enable us to prompt the user before setting a cookie.  While
we only need to prompt before setting, we actually need to make both
CanSetCookie and CanGetCookies asynchronous.  This is necessary in order to
preserve FIFO ordering since the value returned by GetCookies depends on the
changes made to the cookie DB by SetCookie.

This change also includes some simplification of CookieStore.  Instead of N
virtual functions, I distilled it down to only 4.  The remaining functions are
instead expressed in terms of those.

While studying all the places where we currently use CookiePolicy, I found that
some of them were not appropriate.  After discussing with Amit, I decided to
remove the policy checks in URLRequestAutomationJob.  See the comments in the
code regarding this.

I changed the signature of CookieMonster::GetRawCookies to GetAllCookiesForURL
to better match GetAllCookies.  Related to this change webkit/glue/webcookie.h
grows a constructor that takes a CanonicalCookie to help clean up some code.

On the Chrome side, ChromeURLRequestContext now has a ChromeCookiePolicy
object.  That object is threadsafe ref counted because it is passed between the
UI and IO threads.  It is responsible for implementing the queuing logic
described above.  It will also in the future trigger the Chrome UI code to
actually show the setcookie prompt.

Please review the state machinery changes in URLRequestHttpJob carefully.

R=eroman
BUG=34331
TEST=no tests yet for prompting.

Review URL: http://codereview.chromium.org/564045

git-svn-id: http://src.chromium.org/svn/trunk/src/net@38028 4ff67af0-8c30-449e-8e8b-ad334ec8d88c