Refine IPv6 probe to require that the client has an IPv6 address on an interface
This currently only works on Posix, not windows.

Network changes are monitored, and the test is repeated each time interfaces
change (which is a subset of any IP addresses changing).

The test performed is still relatively low latency, and we *may* need
to eventually move to an high latency test, such as a DNS resolution,
or an actual test connection.  If we move in that direction, then we'll
need to post a task to perform the work, rather than immediately returning.

BUG=25680
BUG=12754
r=wtc,eroman
Review URL: http://codereview.chromium.org/652072

git-svn-id: http://src.chromium.org/svn/trunk/src/net@39996 4ff67af0-8c30-449e-8e8b-ad334ec8d88c