Revert 39998 - Revert 39996  Refine IPv6 probe to require that the client has an IPv6 address on an interface
This is a second attempt to land a reviewed change. It was reverted because
the tree got very red (for other reasons), and it was plausible that this
change was causing startup latency in Mac and Linux (causing both perf bots
 to go red).  If this landing turns those perf-bots red (tonight) I'll need
to revert.  (... and I'll need to rearchitect to do the probes
 asynchronously, and get off the startup-critical-path.

This currently only works on Posix, not windows.

Network changes are monitored, and the test is repeated each time interfaces
change (which is a subset of any IP addresses changing).

The test performed is still relatively low latency, and we *may* need
to eventually move to an high latency test, such as a DNS resolution,
or an actual test connection.  If we move in that direction, then we'll
need to post a task to perform the work, rather than immediately returning.

BUG=25680
BUG=12754
r=wtc,eroman
Review URL: http://codereview.chromium.org/652072

TBR=jar@chromium.org
Review URL: http://codereview.chromium.org/660073

TBR=jar@chromium.org
Review URL: http://codereview.chromium.org/661164

git-svn-id: http://src.chromium.org/svn/trunk/src/net@40099 4ff67af0-8c30-449e-8e8b-ad334ec8d88c