A follow-up of r28664.
Remove the unused member completed_handshake_.

In OnHandshakeIOComplete, correct the comments to note
that we also notify the caller of success.  We cannot
DCHECK renegotiating_ is true because DidCompleteRenegotiation
has reset it to false.  Instead, DCHECK the equivalent
conditions in DidCompleteRenegotiation.  Log the end of
TYPE_SSL_CONNECT only for initial handshakes.

In DoCompletedRenegotiation, result may not be OK.

R=mbelshe
BUG=none
TEST=none
Review URL: http://codereview.chromium.org/666004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@40550 4ff67af0-8c30-449e-8e8b-ad334ec8d88c