Revert 42152 - Fix the case where the browser livelocks if we cannot open a file.
If one tries to upload a file that one doesn't have read access to,
the browser livelocks. It tries to read from the file, gets nothing
but spins forever because it knows that it hasn't finished reading.

To address this, firstly we add a check at stat() time to make sure
that we can read the file. However, this doesn't take care of the case
where the access() call was incorrect, or the permissions have changed
under us. In this case, we replace the missing file with NULs.

BUG=30850

Review URL: http://codereview.chromium.org/541022

TBR=vandebo@chromium.org
Review URL: http://codereview.chromium.org/1145004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@42154 4ff67af0-8c30-449e-8e8b-ad334ec8d88c