Remove packet split experiment, and add coalescing histogram
The effort to split the packet proved to not work (presumably 
showing that the RTO was reasonably set by a SYN packet reception).
The test did show that splitting the first packet had a negative
consequence, and so the question is: How often can we coalesce
2 packets in a request into 1 packet?  The most common cause would
be a login, using a POST method to send a body, where the body is 
small.  This change creates a histogram to show percentages of 
requests that can be coalesced.

r=wtc
Review URL: http://codereview.chromium.org/1539003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@43180 4ff67af0-8c30-449e-8e8b-ad334ec8d88c