Fix leak in SpdyNetworkTransactionTest.CloseWithActiveStream.
The test case identified a case where previously we hung (we didn't close out
the stream on the now-dead session).  With the fix, we closed the streams, but
we didn't remove the session reference, leaving it leaked.  The test case did
properly catch this.

BUG=43179
TEST=SpdyNetworkTransactionTest.CloseWithActiveStream.

Review URL: http://codereview.chromium.org/1916004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@46518 4ff67af0-8c30-449e-8e8b-ad334ec8d88c