Reland my close on idle socket change.
This reverts 44402 which reverted r44150 (fix for r43882) and r43882 (original change).  These changes were suspected for causing the infinite loop on the IO thread and also a crash.  It turned out that the backup socket was the cause.  These should be able to be safely relanded.  There's one minor change from the original, due to merging in the weeks of changes in between.
BUG=32817

Review URL: http://codereview.chromium.org/2077004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@47583 4ff67af0-8c30-449e-8e8b-ad334ec8d88c