Revert 54528 - Digest authentication uses a uri field to prevent replay attacks.
[Reason for revert: we want to revert r54505, this was a dependent change.  cbentzel agrees to reland.]

When authenticating to an HTTP proxy to establish a secure tunnel (via CONNECT), the uri should be the hostname of the server and the destination port, such as "www.example.com:443". When authenticating to an HTTP proxy for a non-secure content, the uri should be the path at the server, i.e. "/index.html".

If the site we are trying to connect to previously advertised "Alternate-Protocol: 443:spdy-npn/1" a request to "http://www.example.com" will be attempted on a secure port.
However, the URL passed into the digest authenticator was an unsecure one, and it decided to have a uri in the form "/index.html" rather than the correct "www.example.com:443". This causes persistent failure with the password and many password prompts.

BUG=49865,50822
TEST=Run with --use-spdy=npn, force connection through a digest authenticating proxy, and browse a site which advertises Alternate-Protocol through http URLs.

Review URL: http://codereview.chromium.org/3028021

TBR=cbentzel@chromium.org
Review URL: http://codereview.chromium.org/3091001

git-svn-id: http://src.chromium.org/svn/trunk/src/net@54614 4ff67af0-8c30-449e-8e8b-ad334ec8d88c