Remove test isolation enforcement code.
I think it was a failed experiment, and it only
ever worked for base_unittests.

This removes some potentially risky code from LazyInstance
that was only needed for the test isolation.

The lesson from this experiment is that removing singletons
works better than trying to make tests work fine even when
singletons are overused.

TEST=none
BUG=12710

Review URL: http://codereview.chromium.org/3127012

git-svn-id: http://src.chromium.org/svn/trunk/src/net@56229 4ff67af0-8c30-449e-8e8b-ad334ec8d88c