net: add SSLNonSensitiveHostInfo
Snap Start needs to record a server's certificates and information
about the form of its ServerHello. This CL is part of the work to
support Snap Start in Chrome.

Its intended that a DiskCacheBasedSSLHostInfo is created in
http_cache_transaction.cc and passed into the
http_network_transaction.cc code. There it can be given the the
SSLClientSocket in order to decide if it can do a Snap Start
connection.

BUG=none
TEST=none

http://codereview.chromium.org/3275005/show

git-svn-id: http://src.chromium.org/svn/trunk/src/net@58983 4ff67af0-8c30-449e-8e8b-ad334ec8d88c