Reland r59511: Eagerly set the IO loop used for OCSP.
ChromeOS will create a special Profile for login.  Previously, OCSP initialization was done for the "default" ChromeURLRequestContext for each Profile.  Since we can have multiple profiles, this causes the initialization (and uninitialization) to happen multiple times, which causes problems for OCSP since we use statics.  The solution is to identify the "main" Profile.  We create said Profile in BrowserMain.  I add an "is_main_" variable to URLRequestContextGetter and URLRequestContext, so that only the "main" URLRequestContext will initialize OCSP.

Other than minor plumbing of "is_main_", this changelist is identical to r59511.

BUG=55940
TEST=Startup ChromeOS's Chrome.  Should not crash on startup.

Review URL: http://codereview.chromium.org/3470001

git-svn-id: http://src.chromium.org/svn/trunk/src/net@59972 4ff67af0-8c30-449e-8e8b-ad334ec8d88c