net: remove test flakyness caused by uncorking.
r61546 (reverted r61546, landed r61638, reverted r61672; it was
blameless for the first revert.) caused flakyness in the
SSLClientSocket tests because the SSL_CONNECT event may be hidden by a
SOCKET_SENT event in the case that the test takes over 200ms to run.

This patch changes the test to match against either event.

http://codereview.chromium.org/3531019

BUG=none
TEST=net_unittests

git-svn-id: http://src.chromium.org/svn/trunk/src/net@61956 4ff67af0-8c30-449e-8e8b-ad334ec8d88c