Move the windows-specific scoped_* stuff from base to base/win and in the base::win namespace.
This keeps old headers that forward to the new versions and have using declarations that allow the existing code to compile. I fixed all the callers in base to use the new ones, and also the other files I happened to touch.

This splits out the stuff from scoped_handle into a few separate files. I just deleted ScopedFindFile since it was only used in one place and it wasn't even really helping there.

I removed StackBstr which was a #define and used the "regular" ScopedBstr in the 7 places that used it. This is an optimization to avoid an extra allocation, but none of the callers are remotely performance critical.

TEST=it compiles
BUG=none
Review URL: http://codereview.chromium.org/3781009

git-svn-id: http://src.chromium.org/svn/trunk/src/net@62843 4ff67af0-8c30-449e-8e8b-ad334ec8d88c