A/B test of parelellism limits on DNS resolutions
Current default is a max of 50 parallel resolutions.

This test tries to set limits at 6, 8, 10, 14, and 20.
Stats gathered include the number of successful
resolutions, and well as the number of failed
resolutions.  We're looking to see if the ratio
changes as a function of the parallelism limit.

BUG=3041
r=eroman
Review URL: http://codereview.chromium.org/4010002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@63297 4ff67af0-8c30-449e-8e8b-ad334ec8d88c