Revert 64065 - testserver.py listens on ephemeral ports by default.
I'm not convinced that this caused some of the test failures, but I needed to leave unexpectedly and didn't have time to investigate.

If --port is specified on the command line, testserver.py will listen on that port, otherwise it will listen on an ephemeral port. If --startup_pipe is specified, the port number is written to the pipe as a 2 byte unsigned int in host order.

TestServer by default spawns testserver.py to listen on an ephemeral port and reads the port value from the pipe. If necessary, the port can still be fixed using TestServer::ForcePort, but that will hopefully get deprecated quickly.

BUG=56814
TEST=try bots pass

Review URL: http://codereview.chromium.org/3549003

TBR=cbentzel@chromium.org
Review URL: http://codereview.chromium.org/4165004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@64070 4ff67af0-8c30-449e-8e8b-ad334ec8d88c