snap_start_tests: test NPN misprediction.
When snap startting, we can hit a case where we predicted that we would
negotiation an NPN protocol with a server, but then found that we were
wrong.

r65837 deal with this from the HTTP stack's point of view. This change
adds tests to make sure that we can recover from the misprediction and
get it right next time.

BUG=none
TEST=snap_start_tests

http://codereview.chromium.org/4854002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@66458 4ff67af0-8c30-449e-8e8b-ad334ec8d88c