FTP: improve character encoding detection in cases where ICU's first guess is wrong.
Instead of using ICU's first guessed encoding immediately,
we ask it for all possible encodings, try them in order,
and use the first one that works.

For some sites this still results in a gibberish being displayed,
but at least the links are clickable so the site navigation
is possible.

BUG=61073
TEST=see bug

Review URL: http://codereview.chromium.org/4967001

git-svn-id: http://src.chromium.org/svn/trunk/src/net@66664 4ff67af0-8c30-449e-8e8b-ad334ec8d88c