Allow data URLs to trigger downloads, as they do in Firefox.
Unfortunately, many of our tests rely on being able to use WebURLLoader to load
data: URLs when there is no ResourceLoaderBridge implementation.  In those tests,
we would crash if we try to use the ResourceLoaderBridge.  To workaround that,
and because it probably a good optimization anyways, I decided to check if the
data URL has a supported MIME type, and if it does, then I let it load directly
as before.

Since data URLs may be very large, I modified DataURL::Parse to skip parsing the
'data' section of the URL if the corresponding out param is null.

R=tony
BUG=38546
TEST=none (I would like to add a download test, but they are all disabled or flaky.)

Review URL: http://codereview.chromium.org/5542001

git-svn-id: http://src.chromium.org/svn/trunk/src/net@68212 4ff67af0-8c30-449e-8e8b-ad334ec8d88c