Cleanup X509CertificateMac style nits and introduce a crypto library agnostic means of parsing certificate dates.
In the process of addressing style issues, a bug was found in the date/time parsing routines used on Mac. Though not currently hit, it could lead to wild misrepresentation of the validity period for a certificate. Switch to using the routine used when using OpenSSL, which does things correctly.

BUG=none
TEST=existing

Review URL: http://codereview.chromium.org/4653002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@68322 4ff67af0-8c30-449e-8e8b-ad334ec8d88c