Set PYTHON_PATH at TestServer Init() instead of Start()
Try to look up python paths at Init() time.  It's likely that some paths
in PathService will be overridden later, so if we do not succeed now
we'll try again upon Start().  This is necessary on Chromium OS, as we don't keep the tests next to the browser binary, and InProcessBrowserTest overrides FILE_EXE to point to the browser binary, confusing the test server :-/

BUG=66833
TEST=run browser tests, run chrome os browser test autotest, run ui tests

Review URL: http://codereview.chromium.org/5743008

git-svn-id: http://src.chromium.org/svn/trunk/src/net@70392 4ff67af0-8c30-449e-8e8b-ad334ec8d88c