net: Disable False Start and clear the SSL client auth cache for HTTPS proxies on failure.
When performing TLS client auth with an HTTPS proxy, disable TLS false
start to better handle SSL handshake failures, such as the HTTPS proxy
requiring a client certificate.

In addition, when an HTTPS proxy fails, ensure that it is removed from
the SSL client auth cache, so that if the failure was due to an invalid
client certificate, the user can be prompted to select one again.

Depends on: http://codereview.chromium.org/6017010

Patch by: Ryan Sleevi

BUG=66424
TEST=HttpNetworkTransactionTest.Proxy_ClientAuthCertCache

git-svn-id: http://src.chromium.org/svn/trunk/src/net@71096 4ff67af0-8c30-449e-8e8b-ad334ec8d88c