The SSL server's RSA private key must be imported with the KU_KEY_ENCIPHERMENT key usage to support the RSA key exchange algorithm.
Remove the incorrect workarounds for this bug.  In the
SSLServerSocketTest.DataTransfer unit test, do not proceed to data
transfer if the SSL connection cannot be established.

Not required for fixing this bug: create an RSA private key with all
applicable key usage bits to be future-proof.

R=hclam
BUG=67928
TEST=net_unittests --gtest_filter=SSLServerSocketTest.*
Review URL: http://codereview.chromium.org/6297008

git-svn-id: http://src.chromium.org/svn/trunk/src/net@71739 4ff67af0-8c30-449e-8e8b-ad334ec8d88c