Disk cache: More fixes to avoid corruption.
* We now keep dirty entries in the map of open entries
so that we handle better crashes when we are working with
dirty entries.

* EntryImpl now remembers if the entry was dirty on disk.

* When we find a dirty entry while doing enumerations we
now go through the regular path to delete the entry
(InternalDoomEntry), and let MatchEntry delete the entry.
The main problem with the old code is that it was possible
to delete the entry without first removing it from the index,
so a crash at that time would leave references to free addresses.

* Now we correctly consider the case of not finding an entry
when looking for its parent as an error.

BUG=62085
TEST=net_unittests

Review URL: http://codereview.chromium.org/6538006

git-svn-id: http://src.chromium.org/svn/trunk/src/net@75685 4ff67af0-8c30-449e-8e8b-ad334ec8d88c