FTP: Multiple fixes for localized directory listings:
- fix detection of KOI8-R and possibly other encodings
- fix parsing Russian month names

When detecting the listing encoding, we need to not only
check whether the data can be converted using given encoding,
but also whether the result can be parsed as a valid directory listing.

Also, we only need to compare the first three characters of the
abbreviated month name, because that's how they're abbreviated
in FTP directory listings.

Finally, the Russian directory listings have swapped the "month" and "day of month" columns.

BUG=65917

Review URL: http://codereview.chromium.org/6718043

git-svn-id: http://src.chromium.org/svn/trunk/src/net@81081 4ff67af0-8c30-449e-8e8b-ad334ec8d88c