Revert 89489 - net: don't check revocation when fetching PAC files.
If a PAC file is configured on an HTTPS URL we get into trouble. In order to
check revocation we need to make an HTTP request to the OCSP/CRL server, which
needs the PAC script to load and so we deadlock.

With this change we don't check revocation for PAC fetches.

BUG=86219
TEST=Configure a PAC script on HTTPS.


Review URL: http://codereview.chromium.org/7170026

TBR=agl@chromium.org

NOTE: I'll roll this back in if it turns out unrelated to the ProxyEvents failure.
Review URL: http://codereview.chromium.org/7204023

git-svn-id: http://src.chromium.org/svn/trunk/src/net@89628 4ff67af0-8c30-449e-8e8b-ad334ec8d88c