Speculative fix for bug 80095.
Based off previous CHECKs added, it appears that SpdySessionPool::HasSession() may return true, but later on, may return false. I hypothesize that this is because of HostCache entries expiring in between calls to HasSession().

The solution is not to use HasSession(), but instead just call a new SpdySessionPool::GetIfExists() function. If we acquire it, then cache it in a member variable of HttpStreamFactoryImpl::Job.

BUG=80095
TEST=New tests in HttpNetworkTransactionTest. Because it's timing related, it's hard to come up with a solid repro case.


Review URL: http://codereview.chromium.org/7260014

git-svn-id: http://src.chromium.org/svn/trunk/src/net@90886 4ff67af0-8c30-449e-8e8b-ad334ec8d88c