net: measure the latency difference of resume handshakes.
With False Start, we hope that resume and non-resume handshakes take the same
amount of time. But non-resume handshakes involve sending the certificate chain
to the client and this has a latency impact, although we don't know how high.

There's also a danger that revocation checks will be confounded into this data:
a full handshake is more likely to perform a revocation check. We'll have to
look at the data and, possible, reintroduce the revocation checking field trial
to address this.

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/7218013

git-svn-id: http://src.chromium.org/svn/trunk/src/net@90950 4ff67af0-8c30-449e-8e8b-ad334ec8d88c