Fix instability in SSL client/server sockets
When writing to a socket, net::StreamSocket::Write() may return
immediately, but it may not write the whole buffer passed to it.
In this case SSL sockets were not trying to call Write() again
to send rest of the data, until client tries to call Write()
next time. If client doesn't call Write the remaining data is
never sent to the transport socket.

remoting_unittests was flaky due to this bug.

BUG=88726
TEST=updated net_unittest to catch the bug, remoting_unittests is not flaky

Review URL: http://codereview.chromium.org/7399025

git-svn-id: http://src.chromium.org/svn/trunk/src/net@93445 4ff67af0-8c30-449e-8e8b-ad334ec8d88c