sync: have the client compute position_in_parent values.
This is a protocol change to allow the server to operate on top of
an eventually consistent store.  Rather than having the server do
a lookup of the sibling position to generate an absolute position,
the client will now be responsible for doing that same operation.

As implemented here, math of interpolation is the same as what the
server has historically done.

The client will continue sending the insert_after_item_id, for
interoperation with server implementations that depend on the
old behavior.

Implement the new behavior in the testserver.

TEST=sync_integration_tests,included unittests

Review URL: http://codereview.chromium.org/7481003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@93762 4ff67af0-8c30-449e-8e8b-ad334ec8d88c