On Mac, respect certificate user trust settings created by Safari.
When using Safari to explicitly trust a certificate for a site,
Safari creates the trust settings keyed on the hostname. If the
same trust settings are created in Keychain Access, these settings
are made independent of any particular site.

In order for trust settings to be located, provide the hostname
of the SSL server whose certificate is being validated. This allows
AppleX509TP to locate any Safari-created trust preferences, in
addition to any Keychain-created trust preferences.

BUG=92678
TEST=Manual testing. See bug.

Review URL: http://codereview.chromium.org/7633041

git-svn-id: http://src.chromium.org/svn/trunk/src/net@97842 4ff67af0-8c30-449e-8e8b-ad334ec8d88c