Reland r97821.  DiskCacheBasedSSLHostInfo needs to be declared with NET_EXPORT_PRIVATE for the shared builds.
Do not call callback->Run() until we are about to return from this
object because callback->Run() may cause this object to be deleted.

Previously we were calling callback->Run() in DoWaitForDataReadyDone(),
which returns to the do-while loop in DoLoop().  The do-while loop
reads the state_ member. If the object was deleted in callback->Run(),
we would be reading state_ after free.

Miscellaneous cleanups: omit net::, and use the assign() method of
std::string to avoid constructing a temporary object.

R=agl@chromium.org,rvargas@chromium.org
BUG=84426
TEST=net_unittests --gtest_filter=DiskCacheBasedSSLHostInfo.DeleteInCallback

Review URL: http://codereview.chromium.org/7715007

git-svn-id: http://src.chromium.org/svn/trunk/src/net@98119 4ff67af0-8c30-449e-8e8b-ad334ec8d88c