Always enable the STATE_LOAD_SSL_HOST_INFO state to verify the crash fix for bug 84426.
But we will only use ssl_host_info_->state().certs if the cached info
extension is enabled.

Do not treat an empty ssl_host_info_->state().certs as failure because
it is normal for us to not have the server's certificates if we have
not visited the site before.

R=agl@chromium.org
BUG=84426
TEST=none

Review URL: http://codereview.chromium.org/7741007

git-svn-id: http://src.chromium.org/svn/trunk/src/net@98279 4ff67af0-8c30-449e-8e8b-ad334ec8d88c