[Sync] Add support for enabling session sync remotely.
A "sync_tabs" field has been added to the nigori node. When this field is set,
and if the user has elected to keep everything synced, we automatically enable
session sync for them. This is done by way of the migrator and does not require
a restart of sync or user intervention.

BUG=none
TEST=sync_integration_tests --gtest_filter="*SetSyncTabs*"


Review URL: http://codereview.chromium.org/7669073

git-svn-id: http://src.chromium.org/svn/trunk/src/net@98328 4ff67af0-8c30-449e-8e8b-ad334ec8d88c