NetworkChangeNotifierMac: synchronize destruction order with helper thread.
Previously, it was possible for NetworkChangeNotifierMac to unschedule
the SCNetworkReachability callback before it had been scheduled.  Under the
hypothesis that this is causing crashes on shutdown, this change makes sure
that the helper thread (on which reachability is initially scheduled) runs to
completion before we check to see if we need to unschedule the callback.

BUG=91767

Review URL: http://codereview.chromium.org/7862006

git-svn-id: http://src.chromium.org/svn/trunk/src/net@100771 4ff67af0-8c30-449e-8e8b-ad334ec8d88c