Disk Cache: Improve handling of dirty entries.
* Split the entry sanity checks in two parts: a critical one
and a non-critical one. This allows us to return dirty entries
instead of failing to open them.

* Make sure that we cannot reach an entry through the index
before the actual data reaches the disk (when creating an entry
linked through a parent we were not respecting that).

* When deleting a block from a block file, first clean it
up and then update the map (avoid leaving a dirty free block
if there is a crash)

* Handle the case of errors when opening entries through the
enumerations path.

BUG=73102
TEST=net_unittests
Review URL: http://codereview.chromium.org/8065015

git-svn-id: http://src.chromium.org/svn/trunk/src/net@103323 4ff67af0-8c30-449e-8e8b-ad334ec8d88c