Introduce net::HttpServerPropertiesManager to manage server-specific properties.
Currently the only property we manage is whether or not a server supports SPDY, as indicated by NPN. Also introduce a chrome/ implementation of HttpServerPropertiesManager that persists the information to Prefererences.

When we get a SpdySession for a SPDY server, record that that server supports SPDY in HttpServerPropertiesManager. When preconnecting, if we know that the server supports SPDY, only preconnect 1 socket.

R=willchan
BUG=66472
TEST=browser ui and unit tests,network unit tests

Review URL: http://codereview.chromium.org/7827033

git-svn-id: http://src.chromium.org/svn/trunk/src/net@104666 4ff67af0-8c30-449e-8e8b-ad334ec8d88c