Add WorkerPool::PostTaskAndReply and use in DHCP code.
This factors out the PostTaskAndReply implementation out of MessageLoopProxy so that it can be used for any destination thread accessible via a PostTask-like interface, and uses that code from both MessageLoopProxy and WorkerPool.

The DhcpProxyScriptFetcherWin and DhcpProxyScriptAdapterFetcher classes were both using a PostTaskAndReply-like mechanism with a WorkerPool thread, and on inspection it looks like there are several places in net/ where this is done, and this motivated the larger change (vs. patch set 1 which was a mechanical switch to base::Bind from NewRunnableMethod).

BUG=97516
TEST=net_unittests

Review URL: http://codereview.chromium.org/8139028

git-svn-id: http://src.chromium.org/svn/trunk/src/net@105512 4ff67af0-8c30-449e-8e8b-ad334ec8d88c