Send WM_DEVICECHANGE message through SystemMonitor
WM_DEVICECHANGE is sent when there's been a change to devices or the computer;
specifically when a USB device is connected or disconnected. This is intended
for use in support of Gamepads for more performant polling and
connect/disconnect testing. Currently only on Windows, though seems reasonable
to add for other platforms in the future.

BUG=79050


Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=109960

Review URL: http://codereview.chromium.org/8523021

git-svn-id: http://src.chromium.org/svn/trunk/src/net@110014 4ff67af0-8c30-449e-8e8b-ad334ec8d88c