Use a monotonic clock (TimeTicks) to report network times to WebCore.
Navigation Timing (window.performance.timing) is now spec'd to use monotonically increasing time, so we need to follow suit. Most of this CL is just plumbing TimeTicks through. We already use TimeTicks to implement monotonicallyIncreasingTime() in WebCore.

On Windows, the clock doesn't tick uniformly between processes, so there is a layer in content that corrects for skew.

BUG=None
TEST=Adhoc (ran webtiming-* LayoutTests from Chrome) and content_unittests

Review URL: http://codereview.chromium.org/7602023

git-svn-id: http://src.chromium.org/svn/trunk/src/net@114736 4ff67af0-8c30-449e-8e8b-ad334ec8d88c