base::Bind(): directory_lister.cc - get rid of DataEvent class
Originally, I tried to get rid of the DirectoryLister::Core class to simplify the implementation of DirectoryLister. What I didn't realize was that the purpose of the Core was to outlive the DirectoryLister itself in case the DirectoryLister is destroyed. I think this is probably the most straightforward: Keep the Core, but still make the implementation simpler than it was before we did the base::Bind conversion.

For reference, the previous CL was http://codereview.chromium.org/8989052

BUG=NONE
TEST=visit any file:// url

Review URL: http://codereview.chromium.org/9008059

git-svn-id: http://src.chromium.org/svn/trunk/src/net@116394 4ff67af0-8c30-449e-8e8b-ad334ec8d88c