The network_moved check in DoHandshakeLoop should only apply to STATE_HANDSHAKE.
After transport I/O completes synchronously, stay in the handshake
loop only if the next state is STATE_HANDSHAKE, which is the only
state that requires transport I/O to make progress.  If the next state
is any other state, we should ignore network_moved when determing if we
should stay in the handshake loop.

This changelist is the synchronous completion version of
http://codereview.chromium.org/660131, which dealt with this bug
when transport I/O completes asynchronously.

R=agl@chromium.org
BUG=109706
TEST=Run a Chrome debug build with --enable-origin-bound-certs and
connect to a Google server such as gmail over HTTPS.  Should not hit
a DCHECK failure in net_log.cc.

Review URL: http://codereview.chromium.org/9172005

git-svn-id: http://src.chromium.org/svn/trunk/src/net@117578 4ff67af0-8c30-449e-8e8b-ad334ec8d88c