The network_moved check in DoHandshakeLoop should only apply to STATE_HANDSHAKE.
After transport I/O completes synchronously, stay in the handshake
loop only if the next state is STATE_HANDSHAKE, which is the only
state that requires transport I/O to make progress.  If the next state
is any other state, we should ignore network_moved when determing if we
should stay in the handshake loop.

This changelist is the SSLServerSocketNSS version of
http://codereview.chromium.org/9172005.  SSLServerSocketNSS doesn't
need this change now because it has no other handshake states, but
this change will make it future-proof.

Also remove unnecessary net:: prefixes.

R=agl@chromium.org
BUG=109706
TEST=existing unit tests for SSLServerSocket should continue to pass

Review URL: http://codereview.chromium.org/9250031

git-svn-id: http://src.chromium.org/svn/trunk/src/net@118356 4ff67af0-8c30-449e-8e8b-ad334ec8d88c