Factor out chunk encoding logic into HttpStreamParser::EncodeChunk().
The logic is meaty enough to be factored out.
Add unit tests along the way.

The original patch (crrev.com/118265) was reverted as it introduced
a new static initializer. This version fixed that problem by defining
the constant in the .h file. To be extra careful, replaced
kChunkBufferSize with a member variable chunk_buffer_size_.

BUG=72001
TEST=add unit tests. run tools/linux/dump-static-initializers.py locally to confirm that new static initializers are not introduced.


Review URL: http://codereview.chromium.org/9176009

git-svn-id: http://src.chromium.org/svn/trunk/src/net@118566 4ff67af0-8c30-449e-8e8b-ad334ec8d88c