Adds support for the credentials frame.
Need to allow spdy sessions to see if the underlying socket is required
OBCs. If so, and if pooling a connection, then we need to generate a
CREDENTIALS frame for this origin.

Need to add support for the RST_FRAME status code of INVALID CREDENTIALS

Review URL: http://codereview.chromium.org/9294001

git-svn-id: http://src.chromium.org/svn/trunk/src/net@120923 4ff67af0-8c30-449e-8e8b-ad334ec8d88c