Remove the global HCERTSTORE from Windows in favour of using the NULL HCERTSTORE.
BUG=none
TEST=Existing unit tests should cover all affected functionality.
(Windows Only) On a fresh profile, navigate to different HTTPS sites.
From the Page Info bubble, select Certificate Information, and in the
Windows Certificate Viewer, click "Certification Path" to ensure the entire
chain is displayed. This is a variation of http://crbug.com/45706, which
should not regress.


Review URL: http://codereview.chromium.org/9381016

git-svn-id: http://src.chromium.org/svn/trunk/src/net@122055 4ff67af0-8c30-449e-8e8b-ad334ec8d88c