Add '--host' option to testserver.py and expose it via a new TestServer constructor.
By default the value 127.0.0.1 is used (this is also the default if no option is passed to the testserver.py script). For HTTPS cert mismatch tests, the value localhost is used. This causes the server to listen on localhost/127.0.0.1 but causes the client to expect a cert matching localhost (this corresponds to the previous behaviour).

BUG=114369
TEST=net_unittests still pass

Review URL: http://codereview.chromium.org/9369029

git-svn-id: http://src.chromium.org/svn/trunk/src/net@123040 4ff67af0-8c30-449e-8e8b-ad334ec8d88c