When detecting client certs on OS X, ensure all certs match the server's CertificateRequest
Previously, if an identity preference for a server existed,
such as those created by Safari, then the client certificate
would not be checked to ensure it matched the server's
CertificateRequest. As a result, it was possible to send a
potentially invalid client certificate - and this would
be the default client certificate presented to users.

BUG=118306
TEST=See bug
R=wtc@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9699058

git-svn-id: http://src.chromium.org/svn/trunk/src/net@126836 4ff67af0-8c30-449e-8e8b-ad334ec8d88c