SPDY - Added enabling of SPDY/3  and SPDY flow control to about:flags.
Added command line switches --enable-spdy3 to enable
SPDY/3 and --enable-spdy-flow-control to enable flow control (SPDY/2.1).

Added Field Trials for SPDY/2.1 and SPDY/3. The FieldTrials are enabled
when SPDY/3 and flow control command line options are not chosen.

Removed --use-spdy=v3 and --use-spdy=flow-control options.


BUG=118440,119205
R=willchan
TEST=Netowrk unit tests and browser tests.

Enable SPDY/3 in about:flags and see in chrome://net-internals
that SPDY/2, SPDY/2.1 and SPDY/3 are displayed as the protocols
supported.

Enable SPDY FLow Control in about:flags and see in chrome://net-internals
that SPDY/2 and SPDY/2.1 are displayed as the protocols supported.

Review URL: http://codereview.chromium.org/9766021

git-svn-id: http://src.chromium.org/svn/trunk/src/net@128420 4ff67af0-8c30-449e-8e8b-ad334ec8d88c