Make URLRequestThrottlerManager a member of URLRequestContext.
This addresses a long-standing TODO and allows us to simplify the
logic in the class a bit as it now lives fully on the IO thread.  It
should also allow further cleanup in follow-up changes e.g. to stop
using scoped_refptr for the URLRequestThrottlerEntry instances.

BUG=119760


Review URL: http://codereview.chromium.org/10203002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@134963 4ff67af0-8c30-449e-8e8b-ad334ec8d88c