Fix flakiness and rename ClientSocketPoolBaseTest.DisableCleanupTimer.  Also add a new test where the cleanup timer is disabled, but a socket is reused.
The test used to create two idle sockets that would be timed out after 10 milliseconds, wait 100 milliseconds, run the message loop, and then make sure the sockets were indeed timed out.

The problem is that it would only wait for one of the two sockets to be created before waiting 100 milliseconds.  Sometimes the second socket would be created when the message loop was run a second time (The message loop was being run a second time specifically for the creation of that second socket, actually).  As a result, the socket would be fresh when we expected it to be expired.  Then when we tried to get a new socket, we'd reuse the socket synchronously instead of getting ERR_IO_PENDING, and would hang when waiting for a new socket to be created.

This fixes the issue by getting rid of the 10 millisecond delay before sockets time out.  It also reduces the 100 millisecond delay down to 10 milliseconds, since it's no longer needed.  It also adds some asserts so that rather than hanging, the test should just fail on regression.

R=willchan@chromium.org
BUG=126112
TEST=ClientSocketPoolBaseTest.DisableCleanupTimerReuse,
     ClientSocketPoolBaseTest.DisableCleanupTimerNoReuse

Review URL: http://codereview.chromium.org/10353019

git-svn-id: http://src.chromium.org/svn/trunk/src/net@135504 4ff67af0-8c30-449e-8e8b-ad334ec8d88c