Revert 138168 - net: Fix a regression that broke file uploading with http authentication
Failing net_unit_tests RequestSocketsMultipleTimesDoesNothing.

When uploading a file with http authentication, UploadData is reused
for a new UploadDataStream. Hence, we should close the files in UploadData
if already opened and read, so we can reread the files from the beginning.

The regression was caused by r123677.  Special thanks to thomas.themel for
identifying the offending patch and helping us to fix the issue.

BUG=128574
TEST=added a unit test; manually confirmed that uploading a file with basic authentication, and multi-round authentication with NTLM and Negotiate worked.

Review URL: https://chromiumcodereview.appspot.com/10408042

TBR=satorux@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10388221

git-svn-id: http://src.chromium.org/svn/trunk/src/net@138175 4ff67af0-8c30-449e-8e8b-ad334ec8d88c