[sync] Add --port and --xmpp-port parameters to run_testserver.cc
run_testserver is a useful test utility that launches a python
testserver after setting the required python path.

When it is used to start a sync testserver, the only parameter it
supports is --sync, indicating that we want a sync server. It doesn't
support the --port and --xmpp-port flags, that are used to specify the
ports on which we want the sync server and sync xmpp server to run.

This patch adds the parameters to run_testserver.cc and plumbs them down
to the TestServer objects created by run_testserver.cc. It also separates out
sync specific testserver code into a new class called LocalSyncTestServer.

Separating the sync server in python code is out of the scope of this CL
because it requires significant refactoring, since it is in use on multiple
platforms, and will be dealt with separately.

BUG= 117559
TEST=Build and run run_testserver with the paramters --sync, --port and --xmpp-port

Review URL: https://chromiumcodereview.appspot.com/10388206

git-svn-id: http://src.chromium.org/svn/trunk/src/net@138878 4ff67af0-8c30-449e-8e8b-ad334ec8d88c