Add netwatcher command-line utility that watches and logs network events
Running netwatcher just creates a NetworkChangeNotifier and a
ProxyConfigService and registers the NetWatcher class as observers
for them.  NetWatcher then logs whenever it receives an event from
the NetworkChangeNotifier or ProxyConfigService.

This is so that developers can quickly get an idea of what events
are triggered when.  It may also be useful for nailing down platform
differences.

Also clean up network_change_notifier.h a bit.

BUG=106034
TEST=


Review URL: https://chromiumcodereview.appspot.com/10578027

git-svn-id: http://src.chromium.org/svn/trunk/src/net@143301 4ff67af0-8c30-449e-8e8b-ad334ec8d88c