NSS: send a fatal alert if the ChannelID callback fails.
By the time that NSS makes a ChannelID callback, we have already committed to
sending a ChannelID on the current connection. Therefore, if we cannot, it's
fatal to the connection.

The only time that we should fail to generate a keypair is if we've been
compiled without ECC support. In that case, we need to ensure that we don't
advertise ChannelID support in the first place.

BUG=127506
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10640007

git-svn-id: http://src.chromium.org/svn/trunk/src/net@143629 4ff67af0-8c30-449e-8e8b-ad334ec8d88c