Refactor Android's NetworckChangeNotifier.
The current implementation of NetworkChangeNotifier on Android
assumes that the embedder has the Android ACCESS_NETWORK_STATE
platform permission, and will crash if it does not.

This change refactors the implementation such that the embedder
pushes state changes into the NetworkChangeNotfier, and makes
the NetworkChangeNotifier a singleton.


Review URL: https://chromiumcodereview.appspot.com/10693068

git-svn-id: http://src.chromium.org/svn/trunk/src/net@146100 4ff67af0-8c30-449e-8e8b-ad334ec8d88c