DiskCacheTest.MappedFile_SyncIO asan fix for Chrome on Android.
Build net_unittests with same defines as net.

Downstream, we discovered an ASAN failure in net_unittests. The
problem is that disk_cache::MappedFile has a member variable
that is conditional on the POSIX_AVOID_MMAP preprocessor macro.

DiskCacheTest.MappedFile_SyncIO test creates an instance of
MappedFile, so it needs to have the same preprocessor macro
defined as in net.

BUG=
TEST=DiskCacheTest.MappedFile_SyncIO


Review URL: https://chromiumcodereview.appspot.com/10692096

git-svn-id: http://src.chromium.org/svn/trunk/src/net@146546 4ff67af0-8c30-449e-8e8b-ad334ec8d88c