Allow any text/* mime type to be rendered inline.
Instead of only allowing text/plain, text/css, and various other
specific MIME types to be rendered inline, allow any MIME type that
starts with text/ with the exception of certain types that don't render
in a human-readable form despite technically being textual. The list of
excepted types is the same as the list in WebKit.
    
WebKit renders all these as plain, preformatted text, which should be
safe. Users can still save the file to disk with "Save Page As" if they
didn't want it rendered, and downloading can still be forced with
Content-Disposition.

BUG=68648

Review URL: https://chromiumcodereview.appspot.com/10694020

git-svn-id: http://src.chromium.org/svn/trunk/src/net@147845 4ff67af0-8c30-449e-8e8b-ad334ec8d88c