[Sync] Add support for performing a GetKey on startup.
The functionality is behind the --sync-keystore-encryption flag, and the key
is not currently consumed by anything, but this lays the groundwork for testing
the server and client interaction.

We request a key anytime we perform a GetUpdates while the cryptographer
does not have a keystore key. But, it is considered an error to request a key
and not receive one, putting us into a state of backoff.

BUG=129665
TEST=sync_unit_tests, running against python server

Review URL: https://chromiumcodereview.appspot.com/10455012

git-svn-id: http://src.chromium.org/svn/trunk/src/net@149248 4ff67af0-8c30-449e-8e8b-ad334ec8d88c