Turn off TLS 1.1.
This change is intended to be merged to the 1180 branch for
Chrome 21 Stable. On the trunk this change is only temporary
and will be replaced by a fix for bug 141629.

R=agl@chromium.org,rsleevi@chromium.org
BUG=142172
TEST=Visit https://www.google.com/ and https://www.facebook.com/.
Click the lock icon. The page info bubble should say
"The connection uses TLS 1.0."

Review URL: https://chromiumcodereview.appspot.com/10828272

git-svn-id: http://src.chromium.org/svn/trunk/src/net@151198 4ff67af0-8c30-449e-8e8b-ad334ec8d88c