Remove unnecessary assignment in transport_client_socket_unittest.cc.
This is part of Chrome for Android upstreaming.

This line was introduced by r137387 which turned |sock| into a scoped_refptr
rather than a raw pointer (that needed to be initialized). This assignment is
obviously not needed anymore.

Internally we don't have this extra '= NULL' which causes an unnecessary diff.

BUG=136750


Review URL: https://chromiumcodereview.appspot.com/10855123

git-svn-id: http://src.chromium.org/svn/trunk/src/net@151276 4ff67af0-8c30-449e-8e8b-ad334ec8d88c