Force python test server output to be unbuffered, so it doesn't mix with gtest output
In browser tests that use a local python server, the python output in the test logs sometimes overlaps with gtest output, resulting in gtest falsely detecting passing tests as incomplete. This is a result of python's default use of buffered output, which gets written to the log file out of order.

This patch forces the python process for local test servers to use unbuffered mode. This way, by the time gtest is ready to log a passing test, all testserver output is already written to the log file.

BUG=147368
TEST= See sync integration test output when it is redirected to a log file, and make sure there are no false negatives.


Review URL: https://chromiumcodereview.appspot.com/10919165

git-svn-id: http://src.chromium.org/svn/trunk/src/net@155587 4ff67af0-8c30-449e-8e8b-ad334ec8d88c