Remove {base,net}_java dependencies from GYP client targets on Android.
On Android, some GYP client targets were explicitly depending on both the
native and Java sides of a library.
This removes the dependencies to {base,net}_java and make the native side
('base' and 'net) of these libraries depend on their Java counterpart.
On Android it rarely makes sense to depend on a single side of a Java/C++
library.
The {base,net}_java can now be considered as "private" targets although GYP
does not support this concept unfortunately (AFAICT).
Note that I made sure that the resulting APKs' size is unchanged.

Additionally, this CL removes 'base_java' (i.e. does not replace it with
'base') from the targets including 'build/apk_test.gypi'. This dependency
should not have been there (in the wrong layer) in the first place. It's needed
by ChromeNativeTestActivity.java which clients should not know about.

BUG=146323
TBR=lipalani,sky,willchan,brettw

Review URL: https://chromiumcodereview.appspot.com/10913083

git-svn-id: http://src.chromium.org/svn/trunk/src/net@155737 4ff67af0-8c30-449e-8e8b-ad334ec8d88c