URLRequestHttpJob::StartTransaction should honour network delegate.
If a network delegate returns net::ERR_BLOCKED_BY_CLIENT inside URLRequestHttpJob::StartTransaction() the job is not cancelled. This does not happen, e.g., when ERR_BLOCKED_BY_CLIENT is returned within URLRequestHttpJob::OnStartCompleted(). This CL attempts to change the behaviour of StartTransaction() to cancel the job on the error signal from the delegate.

BUG=146816
TEST=This is not testable in the browser without custom code changes yet.

Review URL: https://chromiumcodereview.appspot.com/10911151

git-svn-id: http://src.chromium.org/svn/trunk/src/net@158984 4ff67af0-8c30-449e-8e8b-ad334ec8d88c