Removing "WinHttpGetProxyForUrl failed" message.
WinHttpGetProxyForUrl() can fail in some network configurations. For example it often fails when running in a home network when it cannot download the PAC script. In other words it is normal for this function to fail and the calling cade should just be able to handle it (which is does already).

BUG=156135

Review URL: https://codereview.chromium.org/11183007

git-svn-id: http://src.chromium.org/svn/trunk/src/net@162220 4ff67af0-8c30-449e-8e8b-ad334ec8d88c