Sort client certificates by preference before prompting the user
When determining the list of available client certificates, order the returned
client certificates based on heuristics to determine preference. Certificates
are sorted based on validity, including that they're still valid, that
certificates with longer expirations are preferred, and that, barring those,
certificates that are newer are preferred.

BUG=89267
TEST=net_unittests --gtest_filter=X509UtilTest*


Review URL: https://chromiumcodereview.appspot.com/11275167

git-svn-id: http://src.chromium.org/svn/trunk/src/net@166231 4ff67af0-8c30-449e-8e8b-ad334ec8d88c