When ordering client certificates, use a constant time for determining validity.
This avoids possible strict weak ordering issues if a certificate expires
while being evaluated for expiration. The effect of such a violation would be
that an expired certificate might be sorted as more preferable than a valid
certificate, but only if it just expired, and dependent on the number of
certs being compared.

BUG=89267


Review URL: https://chromiumcodereview.appspot.com/11388004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@166588 4ff67af0-8c30-449e-8e8b-ad334ec8d88c