[sync] Add Mobile Bookmarks folder to chromiumsync.py
With https://chromereviews.googleplex.com/4378012, the python sync
server started serving the Mobile Bookmarks folder to mobile chrome
clients. In the past, we couldn't upstream this change to desktop
because we weren't syncing the Mobile Bookmarks folder to desktop
clients.

As of today, we do sync the folder, and therefore, it makes sense to
upstream the change to the desktop version of chromiumsync.py.

This patch adds the Mobile Bookmarks folder to desktop chromiumsync.py
and fixes TriggerCreateSyncedBookmarks to account for the change.

BUG=b/7337603
TEST=sync_integration_tests

Review URL: https://codereview.chromium.org/11358102

git-svn-id: http://src.chromium.org/svn/trunk/src/net@166949 4ff67af0-8c30-449e-8e8b-ad334ec8d88c