FTP: Open a fresh data connection after a command error.
Apparently some FTP servers interpret RFC 959 in a way
that makes it possible for them to close the data connection
on a command error. We can't be sure whether that happened
or not, so just always open a new connection after receiving
an error message.

BUG=121335
TEST=Covered by net_unittests.


Review URL: https://chromiumcodereview.appspot.com/11364224

git-svn-id: http://src.chromium.org/svn/trunk/src/net@167832 4ff67af0-8c30-449e-8e8b-ad334ec8d88c