[net] Disable DnsClient in HostResolverImpl after a streak of consecutive failures.
If a fallback from DnsTask to ProcTask succeeds a number of times in a row, it
indicates that the DnsConfig is likely wrong. Disable DnsClient until the next
DNS change.

Record counts in AsyncDNS.DnsClientEnabled: when DnsClient receives a valid config
(as "true") and when DnsClient is disabled after failures (as "false").

BUG=161199
TEST=net_unittests --gtest_filter=HostResovlerImplDnsTest.DisableDnsClientOnPersistentFailure

Review URL: https://chromiumcodereview.appspot.com/11363219

git-svn-id: http://src.chromium.org/svn/trunk/src/net@167953 4ff67af0-8c30-449e-8e8b-ad334ec8d88c