Track unconditionalizable http requests.
Occasionally the disk cache has a hit for an entry that cannot be used because it is out of freshness, but lacks a validator. Track these in the http request pattern histograms, since these are an interesting type of cached resource (they're in the cache, but useless except for forward/back).

R=rvargas@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11362192

git-svn-id: http://src.chromium.org/svn/trunk/src/net@168411 4ff67af0-8c30-449e-8e8b-ad334ec8d88c