net: add histogram to measure how often we do SSLv3 fallback to renego patched servers.
We would like to do something about a network attacker's ability to trigger
SSLv3 fallback. Yngve from Opera suggested taking the renegotiation extension
as a signal that we should not have fallen back and so this change adds a
histogram to measure how often we see renego when doing a fallback connection.

BUG=none


Review URL: https://chromiumcodereview.appspot.com/11348132

git-svn-id: http://src.chromium.org/svn/trunk/src/net@169531 4ff67af0-8c30-449e-8e8b-ad334ec8d88c