Add IncludeForRequestURL method to CanonicalCookie.
Move the code for checking whether to include a cookie or not from the CookieMonster to the new method in CanonicalCookie. This means the new method IncludeForRequestURL returns true when the cookie should be included for the given |url| and cookie options.
Change the CookieMonster to use the IncludeForRequestURL method of CanonicalCookie to test whether a cookie should be included or not.

TEST=CanonicalCookieTest.IncludeForRequestURL  (net_unittests)
BUG=none

Review URL: https://chromiumcodereview.appspot.com/11308272

git-svn-id: http://src.chromium.org/svn/trunk/src/net@172668 4ff67af0-8c30-449e-8e8b-ad334ec8d88c