[Sync] Add support for keystore key rotation.
Key rotation will trigger a full re-encryption of all sync data with the newest
keystore key. Previous keys will be added to the keybag as well. We detect key
rotation by checking whether we have multiple keystore keys and the nigori's
keybag is not encrypted with the current keystore key. In that case, we no
longer support backwards compatibility with non-keystore supporting versions,
and re-encrypt using the newest keystore key.

This change also fixes two latent issues: lack of encryption of the keystore
bootstrap and not properly posting OnPassphraseAccepted when we resolve pending
keys.

BUG=163744


Review URL: https://chromiumcodereview.appspot.com/11434070

git-svn-id: http://src.chromium.org/svn/trunk/src/net@173830 4ff67af0-8c30-449e-8e8b-ad334ec8d88c