Track occurance of Open/Create races on http cache transactions.
If two requests for the same uncached resource occur contemporaneously, then both can attempt to create the same entry. The second will fail to create an entry, and as a result we'll downgrade the request to pass through.

How often does this happen? With this histogram, we'll know.

R=rvargas@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/11711002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@174804 4ff67af0-8c30-449e-8e8b-ad334ec8d88c