Fix miss counting on http cache transactions.
Loads bypassing the cache were counting as misses; this was bad for two reasons. Firstly, we were overcounting misses. Secondly, this was making misses look too performant.

This fix was actually uploaded in https://codereview.chromium.org/11637007/ , but in all the debate about the histograms in that patch, this fix got missed. Here it is again, all by itself.

R=rvargas@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/11741034

git-svn-id: http://src.chromium.org/svn/trunk/src/net@175203 4ff67af0-8c30-449e-8e8b-ad334ec8d88c