Make net/test/test_server correctly report timeouts.
poll() returns 0 on a timeout, and doesn't set errno in that case,
so it makes no sense to use PLOG on a timeout.

Also, this change prints the number of bytes already read, to help distinguish
between a case where we never receive anything from the server, and a case
where we do receive response but not all of it, and the rest gets stuck
for some reason (buffering?).

BUG=96594

Review URL: https://codereview.chromium.org/11784004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@175204 4ff67af0-8c30-449e-8e8b-ad334ec8d88c