Remove support for 'system' SSL on Mac (SecureTransport) and Win (SChannel)
Remove support for the command-line fallback --use-system-ssl, as it
adds to the maintenance burden and presents a security risk due to
inconsistent patching by the OS vendors. On all platforms but Android,
NSS is used for SSL in all situations.

This does not change how certificate verification is handled, just
the underlying SSL implementation.

BUG=170214
TBR=sky

Review URL: https://chromiumcodereview.appspot.com/11876044

git-svn-id: http://src.chromium.org/svn/trunk/src/net@177068 4ff67af0-8c30-449e-8e8b-ad334ec8d88c