Synchronize sending of RST between server and client to avoid dropping bytes.
This is to de-flake the DownloadContentTest.ResumeInterruptedDownload test; 
Windows drops some bytes on the floor in the case of a 0 second linger close, 
which is how we trigger a RST being sent.  To fix this, we do two things:
a) don't send the RST until the client confirms it's received the data, and 
b) make sure to send enough data to defeat the buffering in the download
path.

BUG=169882
R=benjhayden@chromium.org
R=cbentzel@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11968026

git-svn-id: http://src.chromium.org/svn/trunk/src/net@178053 4ff67af0-8c30-449e-8e8b-ad334ec8d88c