Add a command-line flag to have URLFetcher ignore certificate requests
Because having individual URLFetchers ignore certificate
requests affects the URLRequestContext-wide socket pools,
the behaviour is marked as a command-line flag. This is
intended for development purposes when configuring alternate
service endpoints for various URLFetcher-based services that
may be configured to optionally request client auth, and for
which this behaviour cannot be easily turned off.

BUG=168602
TEST=Use --lso-host with a locally-defined service endpoint, ensure that attempts to sign in work.
TBR=thakis


Review URL: https://chromiumcodereview.appspot.com/12095075

git-svn-id: http://src.chromium.org/svn/trunk/src/net@179861 4ff67af0-8c30-449e-8e8b-ad334ec8d88c