Fix a memory error in ProxyResolverV8Tracing.
The problem could happen when executing:
  ProxyResolverV8TracingTest.CancelWhileBlockedInNonBlockingDns

The issue was the Job was not expecting it could get cancelled within the call to the HostResolver. It now handles this case, and I have added a Sleep() so the timing conditions for this problem manifest consistently.

BUG=173373,119151

Review URL: https://chromiumcodereview.appspot.com/12088090

git-svn-id: http://src.chromium.org/svn/trunk/src/net@180021 4ff67af0-8c30-449e-8e8b-ad334ec8d88c