Make the v8 Isolate used in the proxy resolver explicit.
For conceptual and performance reasons, the v8 API will require explicit passing
of Isolates in the future instead of relying on the retrieval of one out of thin
air (= either TLS or an evil global variable in v8). This CL makes the Isolate
explicit in ProxyResolverV8::Context.

The default Isolate is retrieved and remembered in the main browser thread and
then passed down to the Context where it is needed in different threads.

Review URL: https://codereview.chromium.org/11959029
Patch from Sven Panne <svenpanne@chromium.org>.

git-svn-id: http://src.chromium.org/svn/trunk/src/net@180134 4ff67af0-8c30-449e-8e8b-ad334ec8d88c