LoadTiming int net 6:  Cache LoadTimingInfo
LoadTiming information is now cached at the URLRequest
level, so it can be read at any point after the headers
are received, until/unless the request is restarted.

The old behavior of keeping the information only as long as
it could be retrieved resulted in some problems.
Review URL: https://codereview.chromium.org/12114052

git-svn-id: http://src.chromium.org/svn/trunk/src/net@180243 4ff67af0-8c30-449e-8e8b-ad334ec8d88c