SPDY - implement greedy approach to read all the data and process it from the ClientSocket until we block.
This is more of a first cut at the greedy approach to see if it
improves the performance on the mobile.

spdy_session_test_util.* files have the common code between
SpdySessionSpd2Test and SpdySessionSpd3Test. Will move other common
code in a separate CL. Created this file to reduce errors and
to avoid duplicating of code.

R=willchan
BUG=166958
TEST=network unittests

Review URL: https://chromiumcodereview.appspot.com/11644088

git-svn-id: http://src.chromium.org/svn/trunk/src/net@181390 4ff67af0-8c30-449e-8e8b-ad334ec8d88c