Create a new disk_cache type, SHADER_CACHE.
Create a new type of disk_cache to be used by the GL Shader
cache. This cache acts similar to the regular DISK_CACHE
except the rank of the entry is not updated if the entry
is not modified. This means we can read entries as we
iterate and the iterators do not get shuffled around.

BUG=166763


Review URL: https://chromiumcodereview.appspot.com/12224017

git-svn-id: http://src.chromium.org/svn/trunk/src/net@182343 4ff67af0-8c30-449e-8e8b-ad334ec8d88c