When reading from an SSL socket, attempt to fully fill the caller's buffer
The current SSLClientSocket implementation reads one SSL
record at a time, and immediately returns that to the caller
of Read(). As it is a common performance optimization to set
SSL record sizes to fit within MTU, this leads to suboptimal
performance and causes SSLClientSocket::Read() to not
match the behaviour of TCPClientSocket::Read() (which
attempts to fully fill the caller's buffer).

BUG=166903

Review URL: https://chromiumcodereview.appspot.com/12025040

git-svn-id: http://src.chromium.org/svn/trunk/src/net@182912 4ff67af0-8c30-449e-8e8b-ad334ec8d88c