Make synchronous methods on disk_cache::SimpleEntryImpl() reentrant.
For passing browser tests and unit tests, it turns out that the synchronous operations on disk_cache::Entry are called quite a bit. As I was trying to get us to the point that this cache can pass browser tests, I found this was necessary.

Note: this CL is based on https://codereview.chromium.org/12192005/ (Add simple cache backend) , and must land after it.

This CL is the basis of https://codereview.chromium.org/12223075/ (Make Doom asynchronous), and so it must land before that issue.

R=rvargas@chromium.org, felipeg@chromium.org, pasko@chromium.org, willchan@chromium.org
BUG=173392

Review URL: https://codereview.chromium.org/12226095

git-svn-id: http://src.chromium.org/svn/trunk/src/net@184501 4ff67af0-8c30-449e-8e8b-ad334ec8d88c