Added a certificate trust anchor provider.
This provider can return a list of additional certificates to trust when
a trust chain is being built to verify trust in a certain certificate.

This enables the chrome layer to pass in certificates from the ONC policy that
should be trusted only for the user session, without having to import them into
the user's NSS database.

BUG=216495


Review URL: https://chromiumcodereview.appspot.com/12553012

git-svn-id: http://src.chromium.org/svn/trunk/src/net@190070 4ff67af0-8c30-449e-8e8b-ad334ec8d88c