[SPDY] Refactor SpdySession's write queue
This is in preparation for replacing the various IOBuffers used for reads/writes
with a single SpdyBuffer class.

Replace the priority queue of SpdyIOBufferProducers with a SpdyWriteQueue object,
which is an an array of FIFO queues binned by priority.
The priority queue was looking only at priority and so was not guaranteeing
FIFO behavior among producers with the same priority.

Remove the frame queue in SpdyStream and instead have it use the session's
write queue directly.

Remove unused fields from SpdyIOBuffer and clean it up.

Propagate and handle errors from SpdyCredentialBuilder::Build.

Rename SpdyIOBufferProducer to SpdyFrameProducer, have it return a SpdyFrame,
clean up its interface, and move the stream-activating logic out of it.

Replace uses of std::list with std::deque.

Steamline logic in SpdySession that deals with the write queue.

Convert some raw pointers to scoped_ptr<>.

Convert a use of Unretained() in SpdySession to use the weak pointer factory.

BUG=176582

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=192975

Review URL: https://chromiumcodereview.appspot.com/13009012

git-svn-id: http://src.chromium.org/svn/trunk/src/net@194102 4ff67af0-8c30-449e-8e8b-ad334ec8d88c