Avoid temporary std::string creation in net/http/
This patch touches a few functions that take a const std::string&
as a parameter, even though they are called most of the time with
simple raw C strings. The problem is that this forces the
creation/destruction of temporary std::string objects during
the call, which can require heap allocation or memcpy() depending
on the platform.

By using const base::StringPiece& instead, the problem is avoided
since conversion from a C string of std::string is both automatic
and fast.

BUG=233164

Review URL: https://chromiumcodereview.appspot.com/13852014

git-svn-id: http://src.chromium.org/svn/trunk/src/net@195587 4ff67af0-8c30-449e-8e8b-ad334ec8d88c