Add more DiskCacheBackendTest.SimpleCache*.
Every backend test that runs on the memory backend was added. As well,
I cleaned up the order of tests and some names in
backend_unittest.cc. As well, BackendBasics did not pass due to a bug
in the way we handled the index, now fixed.

R=pasko,felipeg,pliard
BUG=None

Review URL: https://chromiumcodereview.appspot.com/14242015

git-svn-id: http://src.chromium.org/svn/trunk/src/net@197656 4ff67af0-8c30-449e-8e8b-ad334ec8d88c