Move FileEnumerator to its own file, do some refactoring.
It creates a class FileInfo to contain the details rather than using a platform-specific typedef. This allows the accessors GetName, GetSize, etc. to be moved directly to this class (previously they were static helpers on the FileEnumerator class) which makes a bunch of code much cleaner. It also gives reasonable getting and initialization which the previous version lacked.

BUG=175002
R=rvargas@chromium.org

Review URL: https://codereview.chromium.org/13165005

git-svn-id: http://src.chromium.org/svn/trunk/src/net@198820 4ff67af0-8c30-449e-8e8b-ad334ec8d88c