Fix NSSCertDatabase Singleton access DCHECK
NSSCertDatabase::ListCerts is being called from a worker thread
in CertLoader. Since NSSCertDatabase is a Singleton, this triggers a
DCHECK when a Debug build is run on a device (see issue).

Simple fix (without re-factoring) is to make NSSCertDatabase a LeakySingleton to avoid the DCHECK.

BUG=241292

Review URL: https://chromiumcodereview.appspot.com/15134007

git-svn-id: http://src.chromium.org/svn/trunk/src/net@201152 4ff67af0-8c30-449e-8e8b-ad334ec8d88c