Rename test callbacks for CookieStore to represent the argument type.
These callbacks are not tightly connected to each method but are made to save
a result value of the specified type. Rename them to represent the type.

Actually, SetCookieCallback is used for testing Delete method which passes
a bool result to the done callback.

Review URL: https://chromiumcodereview.appspot.com/15380003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@201497 4ff67af0-8c30-449e-8e8b-ad334ec8d88c