Land Recent QUIC changes.
Blacklisting PCI domains for quic such that we won't even complete the
crypto handshake for those domains on port 443.

Merge internal change: 46392850

Dont retransmit packets which are NULL encrypted when encryption level
is FORWARD_SECURE.

This solves a bug when the client kept retrying to send a retransmitted
CHLO which the server kept dropping.

Merge internal change: 46380016

QUIC: various cleanups.

* The AES code no longer does key expansion for every packet.
* The CryptoFramer now actually calls Visitor::OnError.

Merge internal change: 46227160

QUIC: add a CryptSecretBoxer object for server-internal needs

Quic[En|De]crypter isn't actually thread-safe. I thought that it was because we
were doing the key schedule every time, but it subtly isn't because the nonce
is stored in the object, not on the stack.

With cl/46095856 (not yet landed) I changed the AES-GCM encrypter/decrypter to
really be non-thread-safe, which seemed fine because it wasn't previously
either, but then I recalled that the server code was assuming otherwise.

Initially I tried adding a thread-safe wrapper that could use thread-local
storage on the server side and just locks in Chrome. That's done in cl/46136485
which is alternative to this CL. However, after writing it, it seemed that it
might be easier just to have a thread-safe boxer, since this is internal to the
server, and be done with it. At the same time we can also solve the short-nonce
worry with the existing, AES-GCM based boxer.

Merge internal change: 46226208

For chromimum: Used sha256 to compute the hash of nonce and plaintext
and stored it at the end of Box message. In Unbox, verified the sha256
hash stored in the ciphertext, matches the sha256 hash of nonce and
plaintext in the ciphertext.

TODO: Use Aes128Gcm12[En|De]crypter instead of sha256.

QUIC - updated the comment in NormalizeHostname to say "stopping at the first trailing dot."

QUIC - moved destructor to out of line.

Merging cleanup changes from chromium - CL 14718011

Merge internal change: 46099778

Adding infrastructure to blacklist domains for QUIC

Merge internal change: 46093221

Changing to be aware of variable length guids, but not yet support reading,
writing, or negotiating non-8 byte guids.

Merge internal change: 46050640

QUIC: invalidate validity when server config changes.

When writing cl/44509151 I started by altering SetServerConfig but changed my
mind and split SetProof from SetServerConfig. I used git to revert the changes
to SetProof to avoid diff noise, but failed to add back the important change:
the validity flag needs to be cleared when the proof is set.

I will be wearing a brown paper bag for the rest of the day.

Merge internal change: 46050043

R=rch@chromium.org

Review URL: https://chromiumcodereview.appspot.com/15910002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@201925 4ff67af0-8c30-449e-8e8b-ad334ec8d88c