Work around GTE CyberTrust/Baltimore CyberTrust cross-signing issues
OS X's lack of robust support for cross-signed certificates, combined with the
impending removal of the legacy GTE CyberTrust 1024-bit root in favour of
the 2048-bit Baltimore CyberTrust Root, will soon cause issues for sites that
need to use the cross-signed intermediate. Fix up the chain on the fly when
dealing with such sites.

BUG=236112
TEST=net_unittests

Review URL: https://chromiumcodereview.appspot.com/14492003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@206274 4ff67af0-8c30-449e-8e8b-ad334ec8d88c