Optimize public suffix list implementation
* Changes the datastructure used for each DomainRule to only have a
char* and an int, instead of also including a bool. The int is then
used as a flag to indicate exception, wildcard and/or private.
* Flipped the order of operations for checking whether this is a private
domain and whether private domains should be included, since most
callers do not include private domains.
* Use -m 10 instead of -m 5 flag when running gperf to further optimize
the size of the lookup table.

BUG=245442
R=pam@chromium.org

Review URL: https://codereview.chromium.org/17155003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@206813 4ff67af0-8c30-449e-8e8b-ad334ec8d88c