Add SocketStreamDispatcherHost::Shutdown and call it when a worker process is being destroyed.
SocketStreamDispatcherHost attached to a worker process should be notified that
the process will be down, before the profile it refers is destroyed.
This CL depends on the current WorkerProcessHost implementation that never add a WorkerInstance
to a WorkerProcessHost once it is created.

SocketStreamHost calls SocketStreamJob::set_context(NULL) which cancels the job.
Since SocketStreamHost::socket_ can be NULL, a NULL pointer guard is added there.

BUG=244746

Review URL: https://chromiumcodereview.appspot.com/16034010

git-svn-id: http://src.chromium.org/svn/trunk/src/net@207218 4ff67af0-8c30-449e-8e8b-ad334ec8d88c