Generates the DTLS identity in browser process and returns it to the renderer process. This is the first part of implementing a persistent identity store for WebRTC in Chrome. It implements the IPC messages to request and return the DTLS identity and the identity generation in browser process. A persistent store based on SQLite DB will be implemented later.
DTLSIdentityService: used in the renderer process to request a DTLS identity. It sends an IPC to the browser process and receives the reply containing the identity in a callback.
DTLSIdentityServiceHost: listens for the IPC message from DTLSIdentityService and passes the request to DTLSIdentityStore.
DTLSIdentityStore: created per RenerProcessHost. For now it always generates a new identity on a worker thread for each request, without any caching or persistent storage.


BUG=https://code.google.com/p/webrtc/issues/detail?id=1864

Review URL: https://chromiumcodereview.appspot.com/15969025

git-svn-id: http://src.chromium.org/svn/trunk/src/net@209499 4ff67af0-8c30-449e-8e8b-ad334ec8d88c