Reland r211790: "Make simple cache not use optimistic operations f..."
The initial attempt failed due to the SimpleAppCacheAppCacheLoad and
SimpleCacheAppCacheEnumerations2 unit tests failing on Mac due to the file
descriptor limit (set to 256).

This CL is needed as part of the default disk backend -> simple backend
migration for app cache (on Android).

This CL also enables a few unit tests in app cache mode with the simple
backend.

BUG=257616

Review URL: https://chromiumcodereview.appspot.com/19369004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@212336 4ff67af0-8c30-449e-8e8b-ad334ec8d88c