Add a HistogramRecorder class and use cases.
This CL adds a utility class to streamline the task of testing that metrics have been updated as expected. Specifically, the HistogramRecorder class allows a client to obtain the differential value of a given histogram in the interval since the given HistogramRecorder instance was created.

BUG=232414

Review URL: https://chromiumcodereview.appspot.com/18337014

git-svn-id: http://src.chromium.org/svn/trunk/src/net@212378 4ff67af0-8c30-449e-8e8b-ad334ec8d88c