Warn users about potentially unwanted downloads.
Exposes POTENTIALLY_UNWANTED ClientDownloadResponse verdict from the
SafeBrowsing service to ChromeDownloadManagerDelegate. Downloads for
which this verdict is seen will be marked with the danger type
DOWNLOAD_DANGER_TYPE_POTENTIALLY_UNWANTED.

A dangerous download prompt will be displayed on the download shelf
indicating that the download may make changes to the browser or
system.

Changes are for Windows only since the POTENTIALLY_UNWANTED signal is
only implemented for Windows.

BUG=262983

Review URL: https://chromiumcodereview.appspot.com/19863005

git-svn-id: http://src.chromium.org/svn/trunk/src/net@213289 4ff67af0-8c30-449e-8e8b-ad334ec8d88c