net: make QUIC ProofVerifier more generic.
The QUIC ProofVerifier code is currently rather Chromium specific: it's using
weak pointers, base::Bind etc. This change wraps abstractions around things so
that the code is more portable again.

It also solves an issue where, when a QUIC connection is canceled while a
verification is running, the verification can write into free memory.

I went back and forth on this change a bit. It effectively reinvents weak
pointers and callbacks in order not to use the Chromium versions of these
things. This is unfortunate but it is desirable to minimise the amount of skew
between different copies of the QUIC code. In the end, the duplicate didn't
seem so bad to me.

Weak pointers are replaced with an explicit callback interface for the
ProofVerifier. The QUIC client stream implements a Cancel method so that it can
cope with being deleted while a proof verification is still running.

The need to store a CertVerifyDetails is taken care of with an abstract class
for wrapping "implementation specific" results from a verification. There is
still Chromium-specific code that casts it to a Chromium object, but that's
unavoidable somewhere. (Although it's not clear that the QUIC client stream is
the best place for it.)

BUG=none

Review URL: https://chromiumcodereview.appspot.com/20047002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@213862 4ff67af0-8c30-449e-8e8b-ad334ec8d88c