Prevent calling PostDelayedTask with negative time deltas.
Handle the calling of Now() more carefully, with consideration of the
fact that Now() can vary from call to call, an hence a comparison with
an "old" value of Now() can not assure how a subtract will work with a
"new" value of Now().

Handle connection_->OnRetransmissionTimeout() returning times that have
already past, such as when it punts on handling all its retransmissions
because it has "done too many" and yeilds.


BUG=264214

Review URL: https://chromiumcodereview.appspot.com/20322002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@213924 4ff67af0-8c30-449e-8e8b-ad334ec8d88c