Add remainder of WebSocketChannel tests
Add test that deleting a failed channel works.

Add a test to verify that deleting the channel from
OnAddChannelResponse() when it indicates failure works as advertised.

Add PingAfterClosingIsRejected test.

Ensure OnDropChannel is only called once.

Add tests for delayed writes and 8-bit-cleanness.

Add WebSocketChannel tests WriteFramesOneAtATime and
WaitingMessagesAreBatched which verify the behaviour when SendFrame() is
called when an existing write is pending.

Add test MuxErrorIsNotSentToStream which verifies that status code
kWebSocketMuxErrorSendQuotaViolation is not sent out on the network.

Add tests WrittenBinaryFramesAre8BitClean and
ReadBinaryFramesAre8BitClean which verify that binary frames are
8-bit clean.

SendClose() sends an empty payload for code 1005.

Test for the case that a headerless chunk follows an invalid chunk.

BUG=230756
TEST=net_unittests --gtest_filter='WebSocketChannel*'

Review URL: https://chromiumcodereview.appspot.com/19510008

git-svn-id: http://src.chromium.org/svn/trunk/src/net@213980 4ff67af0-8c30-449e-8e8b-ad334ec8d88c