This CL adds error tracking to SSLPolicy.
For a certificate to be allowed, it must not have any *additional* errors from when it was allowed, but for a certificate to be denied, it need only match *any* errors that caused it to be denied. Denial is checked before allowance.

More specifically:

* If you are checking CertC with ErrorC against the set of saved certificates, we will deny CertC if ErrorC has any overlap with any saved denied certificates. It does not need to be a perfect match.

* If you are checking CertC with ErrorC against the set of saved certificates, we will allow CertC if ErrorC is a perfect match to or subset of any saved allowed certificates.

BUG=263214
R=rsleevi@chromium.org
TBR=abarth@chromium.org

Review URL: https://chromiumcodereview.appspot.com/20931003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@215516 4ff67af0-8c30-449e-8e8b-ad334ec8d88c