Properly handle asynchronous lookups from the ServerBoundCertStore.
r216233 introduced an issue where, upon successful asynchronous completion
of a lookup in the ServerBoundCertStore, the ServerBoundCertService would
return the existing domain-bound cert for the domain to the caller,
then generate a new domain-bound cert for that domain.
Fix it, by returning without generating a new cert upon successful
asynchronous store lookup, and add unit tests to prevent this in the future.

Additionally, r216233 introduced debug checks on shutdown in the asynchronous
store tests, which this fixes.

BUG=259097

Review URL: https://chromiumcodereview.appspot.com/22603004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@216331 4ff67af0-8c30-449e-8e8b-ad334ec8d88c