Change ownership of |job| objects in ClientSocketPoolBaseHelper
In particular, make RemoveConnectJob() transfer ownership of its job to
Group::RemoveJob(), which will then delete it.

Also assert that the group owns that particular job.

This makes it slightly easier to convert this code to using scoped_ptrs,
which in turn makes it easier to modify ClientSocketPoolBaseHelper to
support reprioritization.

BUG=166689
R=rsleevi@chromium.org

Review URL: https://codereview.chromium.org/22999009

git-svn-id: http://src.chromium.org/svn/trunk/src/net@217661 4ff67af0-8c30-449e-8e8b-ad334ec8d88c