[net/dns] Perform A/AAAA queries for AF_UNSPEC resolutions in parallel.
The second DnsTransaction is scheduled as a second job on the resolver's
PrioritizedDispatcher, at the beginning of its priority queue.  The two
DnsTransactions run independently of each other, although if one of them
finishes with an error, the other one will be scrapped immediately, and
the Job will fall back to ProcTask. 

BUG=174992

Review URL: https://chromiumcodereview.appspot.com/19498003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@218616 4ff67af0-8c30-449e-8e8b-ad334ec8d88c