Create WebSocketDispatcherHost and WebSocketHost classes. WebSocketDispatcherHost creates, destroys and routes messages to WebSocketHost objects. WebSocketHost objects own and communicate with net::WebSocketChannel.
Input validation is delegated to net::WebSocketChannel.

Unit tests seem to be unfashionable for these sorts of glue classes, so I haven't created any.

The code to insert WebSocketDispatcherHost as a message filter is not included in this CL, so WebSocket IPCs are still unhandled and this CL changes no behaviour.

There is a known bug in this CL that the WebSocketChannel can be deleted while it is looping through incoming messages, causing a use-after-free error. This will be fixed in a followup CL. See design for fix at https://docs.google.com/document/d/1fMxcEFiVj-H5vmuxhPqsxJAl98GS6_yZ-wqB6j9Wmy0/pub

BUG=301353

Review URL: https://codereview.chromium.org/12637007

git-svn-id: http://src.chromium.org/svn/trunk/src/net@227435 4ff67af0-8c30-449e-8e8b-ad334ec8d88c