[GCM] Initial work to set up directory structure and introduce socket integration
Create new google_api/gcm directory, and add some of the initial base logic
for integrating protobufs with chrome sockets.

A SocketStream is an implementation of the ZeroCopyStream interfaces that allows
for asynchronously retrieving a message (with timeout support) which can then
be parsed by a CodedInputStream, or alternatively allows a CodedOutputStream
to write into a Chrome socket.

Once a SocketStream closes itself (which will happen on error or timeout), the
stream cannot be reused, and a new one should be created (typically with a
socket that has been reconnected). In general, the state of the socket stream
must be checked before interacting with it.

Also fixes issue in SocketTestUtil where mock Sockets aren't taking ownership
of IOBuffers that they're using (exposed by tests).

TBR=darin@chromium.org
BUG=284553

Review URL: https://codereview.chromium.org/23684017

git-svn-id: http://src.chromium.org/svn/trunk/src/net@229533 4ff67af0-8c30-449e-8e8b-ad334ec8d88c