Fix scoping of raw_bytes in URLRequestFileDirJob::OnListFile
data.info.GetName() returns a new FilePath object, and .value() thus
returns a reference to the string within that FilePath object.
This means that the FilePath object can vanish before the reference
to its contents is actually used - meaning whatever happens to be
on the stack will be accessed.
Fix by keeping a copy of the FilePath object in scope across call to
GetDirectoryListingEntry.

Introduced by r205019.

Review URL: https://codereview.chromium.org/48113006

git-svn-id: http://src.chromium.org/svn/trunk/src/net@231821 4ff67af0-8c30-449e-8e8b-ad334ec8d88c