[Net] Add a priority parameter to URLRequest's constructor
This is so that it is clearer what the intended initial priority of
a URLRequest is.

It is also needed so that we can later enforce that if a URLRequest
is set to ignore limits, it has MAXIMUM_PRIORITY; otherwise, we'd have
to mandate that SetPriority() is called before set_load_flags(), which
is fiddly.

Also standardize on a single URLRequest constructor.

BUG=166689
R=jam@chromium.org, jamesr@chromium.org, joth@chromium.org, mmenke@chromium.org, scottbyer@chromium.org, sky@chromium.org, tommi@chromium.org

Review URL: https://codereview.chromium.org/51953002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@232219 4ff67af0-8c30-449e-8e8b-ad334ec8d88c