Fix a crash in the PriorityQueue class which occurs while trying to assign an uninitialized iterator member to a valid iterator instance.
This occurs if we have a valid Pointer instance with a valid iterator being assigned to an empty Pointer instance.
The ListIterator member in this case is uninitialized as per STL conventions and behaves like an uninitialized pointer
leading to crashes when this iterator is assigned to another.

Fix is to have a dummy List class and initialize this iterator to its end.

BUG=none
R=jar@chromium.org, jar

Review URL: https://codereview.chromium.org/75473003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@236076 4ff67af0-8c30-449e-8e8b-ad334ec8d88c