Remove text media types from supported_non_image_types in mime_util.cc
The only user of the list is non_image_map_ and the only user of
non_image_map_ is IsSupportedNonImageMimeType(). This method has code
handling types starting with text/, so types not in
unsupported_text_types doesn't need to be included in
supported_non_image_types.

BUG=none

Review URL: https://codereview.chromium.org/78963005

git-svn-id: http://src.chromium.org/svn/trunk/src/net@236720 4ff67af0-8c30-449e-8e8b-ad334ec8d88c