net: Make HttpNetworkLayer subscribe to power events on Windows.
HttpNetworkLayer unintentionally stopped subscribing for power events 1 year
and 6 months ago (see r86241).

This is not a big problem in practice though since URLRequestJob already
subscribes to power events. The socket pool is also invalidated when the
network change notifier fires a 'no connection' event after a suspend on
desktop.

This makes HttpNetworkLayer subscribe to power events on Windows which is the
platform that initially needed it.

BUG=318326
TEST=see bug
TBR=pauljensen@chromium.org

This CL is just a rebased version of https://codereview.chromium.org/11572039/
(which never landed but was reviewed).

Review URL: https://codereview.chromium.org/89583002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@237508 4ff67af0-8c30-449e-8e8b-ad334ec8d88c