Certificate Transparency: Threading the CT verifier into the SSL client socket.
With this patch, Basic Certificate Transparency status is available to the UI from the cert_status field of SSLInfo.

Please note this patch is based on https://codereview.chromium.org/67513008/ which has been LGTMed but not submitted yet.

Jam@: If you need any details on how it fits with the UI, it's outlined in issue https://codereview.chromium.org/27026002/

BUG=309578

Review URL: https://codereview.chromium.org/76443006

git-svn-id: http://src.chromium.org/svn/trunk/src/net@237785 4ff67af0-8c30-449e-8e8b-ad334ec8d88c