Disabled resource scheduling when using a SPDY proxy
We already disable the artificial delaying of resources when connected
to a SPDY-capable host, this extends that to also include pages that are
served through a SPDY proxy.

As-implemented, as soon as any request for a given client completes and
was served using SPDY over a proxy connection, that client will be
treated as a SPDY-capable client.  Some resources will still be delayed
until the base page (or another request) completes but it is much better
than it was before.

BUG=324789

Review URL: https://codereview.chromium.org/99533002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@243698 4ff67af0-8c30-449e-8e8b-ad334ec8d88c