ChromeOS: Fix crash if login profile triggers client auth.
The login profile (which is identified with an empty username_hash) does not have an entry in the chromeos_user_map_, which would cause a crash (or DCHECK) when GetPrivateSlotForChromeOSUser was called. GetPrivateSlotForChromeOSUser is changed to return a NULL slot handle for this case.

Updates NSSProfileFilterChromeOS to allow NULL slot handles, which it will now receive due to the above change. 

BUG=331945,302125

Review URL: https://codereview.chromium.org/123633002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@244690 4ff67af0-8c30-449e-8e8b-ad334ec8d88c