fetch_client: Support file URL scheme
This is done by moving to URLFetcher rather than HttpNetworkTransaction. This both simplifies the code for fetch_client, and makes it closer to a typical consumer of the networking library.

This CL also fixes URLRequestContextBuilder::DisableHttpCache, which I discovered was broken while making this change.

BUG=335585

Review URL: https://codereview.chromium.org/134603004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@245803 4ff67af0-8c30-449e-8e8b-ad334ec8d88c