On Windows, avoid QUIC UDP-port selection in Stable and Beta channels
We need to avoid instigating security dialogs on Stable or Beta
under Windows.  The port selection feature is helping us appraise
0-RTT connectivity in QUIC by "usually" hitting the same server
when the same destination IP is used (because the source IP and
consistently-selected port will usually be sharded by load balancers
to the same machine each time).

Hopefully, we'll soon have a way to detect that it is safe to use
port selection on Stable and Beta, because the installer will try to
create a security rule allowing port selection.  Until then, we just
decline to use this feature.

r=rch
BUG=329255

Review URL: https://codereview.chromium.org/141803006

git-svn-id: http://src.chromium.org/svn/trunk/src/net@246230 4ff67af0-8c30-449e-8e8b-ad334ec8d88c