Added SSLHostInfo. Storing of server info to our standard disk cache.
Implemented  DiskCacheBasedSSLHostInfo which fetches information about an SSL host from our standard disk cache. Since the information is defined to be non-sensitive, it's ok for us to keep it on disk.

SSLHostInfo persists/restores server's certificates.  Certificates is a vector of DER encoded X.509 certificates, as the server returned them and in the same order.

These changes are the frame work. This code is not enabled to store any data to disk cache.

R=wtc@chromium.org

Review URL: https://codereview.chromium.org/135373002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@246254 4ff67af0-8c30-449e-8e8b-ad334ec8d88c