Add a test for Pong frames with NULL data and fix.
Previously, passing a frame with NULL data to
WebSocketBasicStream::WriteFrames() could cause a crash.

This is a particular problem because WebSocketFrameParser creates frames with
NULL data when there is no payload. This is particularly common with Ping
frames. WebSocketChannel copies the original payload from a Ping message to the
Pong message, so if it is NULL it stays NULL.

Stop attempting to access the data for zero-size payload messages in
WebSocketBasicStream.

It appears that the tests for WebSocketDeflateStream already use NULL for empty
payloads and it is not vulnerable to this problem.

BUG=337868
TEST=net_unittests --gtest_filter=WebSocket*

Review URL: https://codereview.chromium.org/144643005

git-svn-id: http://src.chromium.org/svn/trunk/src/net@247210 4ff67af0-8c30-449e-8e8b-ad334ec8d88c