Use endianness detection from build/build_config.h
The file build/build_config.h already do the endianness detection
based on the cpu, so remove duplication by directly using the macro
ARCH_CPU_{LITTLE,BIG}_ENDIAN to detect endiannes.

This is required to fix the compilation to arm64 for iOS as it does
not define the ARCH_CPU_ARMEL macro (the instruction set is not
compatible between armv7 and arm64).

BUG=314720
R=wtc@chromium.org

Review URL: https://codereview.chromium.org/135123015

git-svn-id: http://src.chromium.org/svn/trunk/src/net@250415 4ff67af0-8c30-449e-8e8b-ad334ec8d88c