Remove failing CHECK(!delegate_) from SpdyStream::OnDataReceived()
There is a race between SetDelegate() and the PushedStreamReplay() call
it posts a task for, and OnDataReceived() called as more data is
received over the network.  Specifically, if the latter is called after
SetDelegate() is called but before the task it posted has been executed,
the CHECK() fails.

BUG=356110

Review URL: https://codereview.chromium.org/210733006

git-svn-id: http://src.chromium.org/svn/trunk/src/net@259800 4ff67af0-8c30-449e-8e8b-ad334ec8d88c