Backout SpdySettingsID implicit renumbering.
Renumbering was introduced in https://codereview.chromium.org/202383002/
SpdyConstants (added in that CL) is now responsible for mapping settings
to protocol-specific wire values. SpdySettingsIds was absolved of
responsibility for capturing wire settings ID's, and the removal of
explicit enum values was cosmetic.

However it's interacting poorly with settings persistence, which is caching
setting ID's keyed on the enumeration values. So back it out.

BUG=358677

Review URL: https://codereview.chromium.org/221173004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@261021 4ff67af0-8c30-449e-8e8b-ad334ec8d88c