Add experiment to measure time to hash extension content as we read it
This required adding 2 new methods to URLRequestFileJob to let
subclasses get notified when seeks and reads complete, so that our
URLRequestExtensionJob class is able to compute the hash of the content
read and know what bytes from the file those correspond to (e.g. if there
was a Range specified and we didn't read all the bytes).

BUG=360909

Review URL: https://codereview.chromium.org/227943003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@263571 4ff67af0-8c30-449e-8e8b-ad334ec8d88c