Disk cache: Remove stats_histogram.
StatsHistograms is the only "external" subclass of base::Histogram and as such
it often requires extra considerations when modifying the base class. On the
other hand, almost the same data can be gathered by using a regular log
histogram with a specific range.

This CL does that, generating a 75 bucket histogram to map the 27 buckets
from the original data. Most of the buckets will be empty, and a couple of
buckets will be merged on the new histogram, but the compromise looks
good enough.

BUG=377936

Review URL: https://codereview.chromium.org/295913014

git-svn-id: http://src.chromium.org/svn/trunk/src/net@273074 4ff67af0-8c30-449e-8e8b-ad334ec8d88c