Support recovery from SSL errors.
Previously, the new WebSocket implementation was unable to handle sites
with self-signed certificates and other cases where the user had
overridden certificate errors. Add code to support this case.

This requires adding infrastructure to pass the SSL error back up to the
content layer which knows how to handle it. It also requires that the ID
of the frame be known, so an extra parameter has been added to the
WebSocketHostMsg_AddChannelRequest IPC message.

BUG=364361

Review URL: https://codereview.chromium.org/304093003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@275066 4ff67af0-8c30-449e-8e8b-ad334ec8d88c