Componentize URLFixerUpper.
This CL moves URLFixerUpper into a new url_fixer component to allow this code
to be shared by the iOS port. At the present time the component contains some
minor Chrome-specific logic. Specifically, it rewrites about:// to chrome://
and uses about:// version as the default about:// host. We decided not to
abstract this logic at this time, as the only embedders wishing to use this
component are ports of Chrome.  However, if there comes to be a non-Chrome
embedder that wishes to use this component, this behavior could easily be
generalized (e.g., by exposing the variables that are used for these purposes
in the header file to allow the embedder to customize them).

BUG=373229
R=jam@chromium.org
TBR=mmenke
NOTREECHECKS=true

Review URL: https://codereview.chromium.org/320253004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@276672 4ff67af0-8c30-449e-8e8b-ad334ec8d88c