Mojo: Load document data through WebURLLoader.
Implemented by setting the DataPipeConsumerHandle and URLResponse object as "extra data" on the WebURLRequest passed to WebFrame::loadRequest.  This enables Blink to think that it is doing a normal navigation request. It also allows us to stream the data to the HTML parser rather than buffering it all in memory up front.

It was necessary to plumb a few more response fields through from the network stack, in particular the mime type. This is expected by Blink's DocumentLoader.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/344413002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@279272 4ff67af0-8c30-449e-8e8b-ad334ec8d88c