net: only consider CRLSet coverage for the leaf certificate.
Currently, CRLSet coverage requires that every certificate in a chain be
covered. However, some intermediates have CRLs with no revocations and those
CRLs are pruned from the CRLSet at generation time. This means that
some EV sites are taking the hit of an OCSP lookup for no reason.

We could include empty CRLs in the CRLSet, but being able to prune them is
nice. Instead, this change redefines coverage so that only the leaf certificate
is considered on the basis that revoking an intermediate in anger is
sufficiently rare and important that we'll be taking extraordinary measures in
any case.

The OS X code isn't changed because it doesn't consider CRLSet coverage and (I
think) always does OCSP lookups for EV certs due to platform limitations. The
OpenSSL and Android code isn't changed because it doesn't implement CRLSet
checking.

BUG=none

Review URL: https://codereview.chromium.org/342033003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@279521 4ff67af0-8c30-449e-8e8b-ad334ec8d88c