Add header pointer to BeforeProxyHeadersSentCallback
This is needed because the higher layers, e.g., UrlRequestHttpJob
do not always have a pointer to the request headers that the
HttpNetworkTransaction will actually send. Note that the old code
could also pass the wrong headers if redirects resulted in no longer
going through the proxy.

BUG=345907

Review URL: https://codereview.chromium.org/360113004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@280885 4ff67af0-8c30-449e-8e8b-ad334ec8d88c