Implement TLS_FALLBACK_SCSV for SSLClientSocketOpenSSL.
In doing so, fix a bug in tlslite's TLS_FALLBACK_SCSV support; the fallback
alert should be sent with the client's version. Otherwise OpenSSL reports
SSL_R_UNSUPPORTED_PROTOCOL and doesn't report the alert. This behavior is
probably not wrong as, if the server responds with a TLS version higher than
what is supported, we can't really be sure of the parse.

BUG=388425

Review URL: https://codereview.chromium.org/418553003

git-svn-id: http://src.chromium.org/svn/trunk/src/net@285764 4ff67af0-8c30-449e-8e8b-ad334ec8d88c