Attempt to fix UnixDomainServerSocketTest.AcceptWithForbiddenUser under Valgrind on Mac.
The issue is most likely the assumption the test makes that
running the message loop is enough for a socket message to make it
from the server socket to the client socket.  The fix tries to
read from the client socket instead, and waits for the result.

BUG=397031

Review URL: https://codereview.chromium.org/420823002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@286911 4ff67af0-8c30-449e-8e8b-ad334ec8d88c