Escape non-ASCII response headers.
Response headers that are not valid UTF-8 can't be stored in a
base::StringValue, which is always a UTF-8 string, so escape any non-ASCII
values to be safe. This prevents a DCHECK when dev tools' network request
inspector is open and we receive these headers.

BUG=299880

Review URL: https://codereview.chromium.org/453483002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@288357 4ff67af0-8c30-449e-8e8b-ad334ec8d88c