Break labels apart in preloaded list.
Since the hosts are stored in DNS wire format (with length-prefixed labels),
any labels that start with a digit cause an escape like "\0071..." (that's
length=7, first byte of the label = "1"). But this causes a warning in Visual
Studio and breaks the build.

This change splits the length bytes from the labels so that this warning
doesn't happen.

BUG=none
R=palmer@chromium.org

Review URL: https://codereview.chromium.org/456853002

git-svn-id: http://src.chromium.org/svn/trunk/src/net@288443 4ff67af0-8c30-449e-8e8b-ad334ec8d88c