Export URLRequestContextBuilder::SchemeFactory
The inner class needs to be exported in case any uses in inline methods
or vtable of the outer class causes it to potentially be referenced across
a shared library boundary.

On non-Windows, the visibility attribute is already inherited by nested
classes, but on Windows, they must be marked explicitly.

BUG=82385

Review URL: https://codereview.chromium.org/464563004

git-svn-id: http://src.chromium.org/svn/trunk/src/net@289093 4ff67af0-8c30-449e-8e8b-ad334ec8d88c