Move StringToUpperASCII and LowerCaseEqualsASCII to the base namespace
Convert LowerCaseEqualsASCII to take StringPiece. In the current patch this is
generally a NOP but will allow me to delete the other 4 variants in a followup
(wanted to do that separately since that will require more review, since
callsites will be changed in nontrivial ways).

In some cases, LowerCaseEqualsASCII is called with a WebString, which no
longer is implicitly converted. I added base::string16(...) around such
calls to force the right conversion. It happened in these files:
  window_container_type.cc
  savable_resources.cc
  render_view_impl.cc
  blink_ax_tree_source.cc
  password_form_conversion_utils.cc
  translate_helper.cc
  chrome_render_view_observer.cc
  dom_serializer_browsertest.cc

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/448143008

git-svn-id: http://src.chromium.org/svn/trunk/src/net@289312 4ff67af0-8c30-449e-8e8b-ad334ec8d88c