Http Cache: Relax DCHECK to account for an entry waiting on the message loop.
HttpCache::DoomEntry has a DCHECK to verify that the ActiveEntry being doomed
looks fine. However, if the entry is being closed because the writer is done
and there are no readers, there is a small window when that DCHECK will fire,
while the task to process the pending queue goes through the message loop.

This CL accounts for that window.

BUG=404607
R=michaeln@chromium.org

Review URL: https://codereview.chromium.org/486873006

git-svn-id: http://src.chromium.org/svn/trunk/src/net@290438 4ff67af0-8c30-449e-8e8b-ad334ec8d88c