Log SPDY protocol specific setting IDs (alternate).
This is an alternate way of achieving the same as https://crrev.com/568093002.
In this version, instead of calling ToEventParametersCallback to get a
callable, then passing it around, then calling it, the body of
SourceEventParametersCallback is duplicated in NetLogSpdyInitializedCallback.
This yields faster, simpler, more direct code, at the expense of code
duplication.

Only one of this CL and https://crrev.com/568093002 should be committed.

BUG=413369

Review URL: https://codereview.chromium.org/567403002

Cr-Original-Commit-Position: refs/heads/master@{#295059}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 382003433672bca0ac734299c91dfe14ec167f05