Add debugging information to the SSL blocking page
If you click on the SSL interstitial's error code, the following is
now shown:
* subject
* issuer
* validity end date
* current date
* pem encoded chain

This CL adds a placeholder for debugging information to the security
interstitial. It's intentionally meant to be flexible, so that it could
support other debugging information in the future from other warnings.

BUG=413838

Review URL: https://codereview.chromium.org/584463005

Cr-Original-Commit-Position: refs/heads/master@{#296349}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: df2931a70908043e07de24d3fa48c25cb68d7dc1