Unescape BiDi control chars while parsing data: urls
The fix for bug 337746 prevented unescaping of BiDi control
characters in URLs. This breaks the loading of data: URLs
because BiDi control chars appear escaped in the loaded HTML.
This patch adds a special case for the parsing of BiDi control
chars. This shouldn't change the way URLs are shown in the omnibox
or any other UI. URLs with BiDi control characters are always
displayed as escaped in the omnibox. This behavior is also
consistent with Firefox.

BUG=423901

Review URL: https://codereview.chromium.org/643963004

Cr-Original-Commit-Position: refs/heads/master@{#300584}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c62e2eeb225afbc123db27c2d3500a7bbbab6a48