Update EscapeExternalHandlerValue to keep '#', '[', and ']'
According to RFC 3986, '#' is used as a separator of fragment part, and
'[' and ']' are used for IPv6 address.
So they should be kept as they are in a URL string.
And EscapeExternalHandlerValue should escape '%' if it isn't used as a
prefix for %XX.

BUG=419257

Review URL: https://codereview.chromium.org/646333002

Cr-Original-Commit-Position: refs/heads/master@{#301013}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9e212e96140c2cfc9d259ef6ce620895d0e5a384