Check effective proxy config before adding data reduction proxies
Before this change, the data reduction proxies were added to the
list of acceptable proxies for a URL, if the effective proxy
configuration resolved to direct for it. This change further checks
that the effective configuration is not the data reduction proxy's.

BUG=425826

Review URL: https://codereview.chromium.org/653313003

Cr-Original-Commit-Position: refs/heads/master@{#302562}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f5c403d7f2a789f755428505231861988f9c40fd