do not send traffic to localhost through proxy for android
The proxy config service for android does not allow
one to configure an exception for localhost be made.
The implementation for Linux did allow this in:
   proxy_config_service_linux.cc
it just seems it was overlooked for android.

Instead of adding in the ability to configure
localhost as an exclusion, it makes more sense
to just force the exception; it seems silly to
send traffic for localhost through a proxy.

This patch addresses the following bug in android:
   https://code.google.com/p/android/issues/detail?id=37031

BUG=

Review URL: https://codereview.chromium.org/796003002

Cr-Original-Commit-Position: refs/heads/master@{#308243}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 80295948bf0bc43ebb640afd81fa152e9679fd36