Http Cache: Improve handling of unexpected responses with byte ranges.
The main change here is to keep track of the end of the requested range
so that when a server returns a different range, we can make sure we
detect the condition and update the response accordingly. Previously we
could detect the condition but fail to update the returned headers (or
more preciselly, overwrite the range header as if nothing happen).

The second change is to avoid reissuing the same network requests twice.
When we detect an unexpected response or an error, we often try to
remove headers changed by the cache an reissue the request; This could
cause the same request issued again (if we didn't add anything to start
with). Now we avoid reissuing requests if there was nothing cached to
start with (so we did not have to modify the requested range).

BUG=421787, 421266
TEST=net_unittests

Review URL: https://codereview.chromium.org/739803002

Cr-Original-Commit-Position: refs/heads/master@{#310015}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3b57e37a2cd304a7b191c245a35f15274629fd73