net: Don't store byte-range resources that cannot be revalidated.
This is basically a revert of
https://crrev.com/439a05cd87fe1f880b896bd7b154454299355899
Cr-Commit-Position: refs/heads/master@{#292751}

The main purpose of said change was to enable code to improve seek operations
when the server is not well configured (think videos without the proper
caching headers). That requires at least a load flag to specify the intent.
However, we can wait until there's actually an attempt to implement that code
before we start saving everything.

The downside from saving these resources is that they are basically unused so
they just take space and processing time. They could be used if they have an
expiration date but lack validators, but that combination is probably very
uncommon.

BUG=407923, 442318
TEST=net_unittests
R=davidben@chromium.org

Review URL: https://codereview.chromium.org/834213003

Cr-Original-Commit-Position: refs/heads/master@{#310406}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 43dc8fd3842f24de52afdc01cad25a500570caa5