Do not advertise h2-15 support.
Stop advertising h2-15 on trunk so that Beta and Stable will not do so.
Intended for M41, might need to be merged.  Can be reverted later if h2-15
support is needed on Dev and Canary for testing.

BUG=447587

Review URL: https://codereview.chromium.org/822153003

Cr-Original-Commit-Position: refs/heads/master@{#310854}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 34ffac8d3c9de489b99d1e3c8ba968284af98706