Adding instrumentation to locate the source of jankiness.
Mechanical change that adds instrumentation required to locate the source of
jankiness (i.e. a long-running fragment of code executed as a part of the task
that causes jank) in the code. See the bug for details on what kind of jank we
are after.  A number of similar CLs were landed, and none of them caused issues.
The code of the instrumentation is highly optimized and is not expected to
affect performance. The code simply creates a diagnostic task which is identical
to ones created by PostTask or IPC message handlers.

BUG=456327,456331,456373
TEST=none
TBR=mmenke, erikwright

Review URL: https://codereview.chromium.org/894903004

Cr-Original-Commit-Position: refs/heads/master@{#315173}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b60049a86e45b0160f378624c5dfd99d667c5262