Use int64 time stamp when storing metadata to the HTTP cache.
This is part 2/4 of a clean up to switch to int64 time stamps when caching metadata. The background to this is that converting int64 time stamps to double and back to int64 is lossy and causes false time stamp mismatches.

BUG=453298

Review URL: https://codereview.chromium.org/935963002

Cr-Original-Commit-Position: refs/heads/master@{#317421}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7fbf4c1ca967c4dd4674622f2c2ea059d08566b7