First-Party Cookies: Wire it up as an experimental web platform feature
This patch adds a flag to the NetworkDelegate to control the first-party
cookies experiment, and implements the flag in the ChromeNetworkDelegate
and ShellNetworkDelegate by checking the value of the experimental web
platform features command-line flag.

Once we decide whether or not to ship this feature, we can revert
everything in this patch other than the tests and the change to
URLRequestHttpJob::DoLoadCookies.

BUG=459154

Review URL: https://codereview.chromium.org/940373002

Cr-Original-Commit-Position: refs/heads/master@{#318295}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3f3daac24caec2913641fe2fc2685feb738676ad