Simplify GetPyProtoPath.
We shouldn't hard code out/Release (since with GN the path could be different) or look for the checkout directory (since with swarming it doesn't exist).

This was needed with ui_tests or pyauto, but not with the current browser_tests binaries.

BUG=463180

Review URL: https://codereview.chromium.org/969313002

Cr-Original-Commit-Position: refs/heads/master@{#318892}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 35e1c076cbda56bd384f1b409f1bd782bd0fb70d