Disk cache: Re-initialize stats counters if they are zero on-disk.
Stats are saved to disk every five minutes. If a client is initializing
the cache, and it crashes before saving the stats for the first time,
the next run will see an empty stats storage, and will discard the cache
because it may indicate that the storage address was invalid.

Now, we check the contents to see if the storage was empty and re-initialize
things as needed.

BUG=none
TEST=net_unittests

Review URL: https://codereview.chromium.org/980003002

Cr-Original-Commit-Position: refs/heads/master@{#319204}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f2eab7cd613d87c79a192bfc5183f79483d95bf