net:  Fix a NULL dereference when ICU fails to create an exemplar set.
IsComponentCoveredByLang wasn't checking if ulocdata_getExemplarSet fails,
as it currently does for the bn, gu, and te language codes.  This CL adds
a check for that.

The downside of this is that we won't decode IDN hostnames consisting
solely of characters in those sets.

It's unclear if ulocdata_getExemplarSet failing is an ICU bug, a CLDR bug,
or expected behavior, but the docs allow the method to fail, and if it does,
Chrome shouldn't crash.

BUG=410551

Review URL: https://codereview.chromium.org/973203003

Cr-Original-Commit-Position: refs/heads/master@{#319299}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 187f371951a9f3d55a3db89b1f8e9c8ebf10d8db